+++ R/plot.R
@@ -367,8 +367,8 @@ genPlotGIS<-function(r,region,breaks,labels,zlim,layout,proj,nbreaks=40,nlabels=
   if(inherits(r,"list")){
     if(inherits(r[[1]],"RasterBrick")||inherits(r[[1]],"RasterStack")||inherits(r,"RasterLayer")){
       if(!missing(proj)){
-        r = lapply(r, projectRaster,crs=proj)
-        if(!missing(region)){region=transform_multiple_proj(region,proj4=projection(r[[1]]))}
+        r <- lapply(r, projectRaster,crs=proj)
+        if(!missing(region)){region<-transform_multiple_proj(region,proj4=projection(r[[1]]))}
       }
     }else if(inherits(r[[1]],"stars")){
 
@@ -377,8 +377,8 @@ genPlotGIS<-function(r,region,breaks,labels,zlim,layout,proj,nbreaks=40,nlabels=
     }
   }else if(inherits(r,"RasterBrick")||inherits(r,"RasterStack")||inherits(r,"RasterLayer")){
     if(!missing(proj)){
-      r = projectRaster(r,crs=proj)
-      if(!missing(region)){region=transform_multiple_proj(region,proj4=projection(r))}
+      r <- projectRaster(r,crs=proj)
+      if(!missing(region)){region<-transform_multiple_proj(region,proj4=projection(r))}
     }
   }else if(inherits(r,"stars")){
 
@@ -391,25 +391,25 @@ genPlotGIS<-function(r,region,breaks,labels,zlim,layout,proj,nbreaks=40,nlabels=
   # layout preconfigured arguments
   tm_layout_args<-args[names(args)%in%names(formals(tm_layout))]
   if(!("legend.bg.color" %in% names(tm_layout_args))){
-    tm_layout_args$legend.bg.color="white"
+    tm_layout_args$legend.bg.color<-"white"
   }
   if(!("panel.show" %in% names(tm_layout_args))){
-    tm_layout_args$panel.show=TRUE
+    tm_layout_args$panel.show<-TRUE
   }
   if(!("panel.labels" %in% names(tm_layout_args))){
-    tm_layout_args$panel.labels=names(r)
+    tm_layout_args$panel.labels<-names(r)
   }
   if(!("legend.outside" %in% names(tm_layout_args))){
-    tm_layout_args$legend.outside=TRUE
+    tm_layout_args$legend.outside<-TRUE
   }
   if(!("legend.outside.size" %in% names(tm_layout_args))){
-    tm_layout_args$legend.outside.size=0.08
+    tm_layout_args$legend.outside.size<-0.08
   }
   if(!("legend.outside.position" %in% names(tm_layout_args))){
-    tm_layout_args$legend.outside.position="right"
+    tm_layout_args$legend.outside.position<-"right"
   }
   if(!("frame" %in% names(tm_layout_args))){
-    tm_layout_args$frame=TRUE
+    tm_layout_args$frame<-TRUE
   }
 
   if(!missing(layout)){
@@ -423,15 +423,15 @@ genPlotGIS<-function(r,region,breaks,labels,zlim,layout,proj,nbreaks=40,nlabels=
   tm_graticules_args<-args[names(args)%in%names(graticules_args)]
   names(tm_graticules_args)<-graticules_args[names(tm_graticules_args)]
   if(!("lines" %in% names(tm_graticules_args))){
-    tm_graticules_args$lines=FALSE
+    tm_graticules_args$lines<-FALSE
   }
   if(as.grid){
-    tm_layout_args$between.margin=-.1
[…]
