+++ R/cloud_mask.R
@@ -11,120 +11,150 @@
 #' @include rtoi.R
 #' @export
 #' @importFrom zip zip_list
-setGeneric("cloud_mask", function(x,...) {standardGeneric("cloud_mask")})
+setGeneric("cloud_mask", function(x, ...) {
+  standardGeneric("cloud_mask")
+})
 
 #' @rdname cloud_mask
 #' @aliases cloud_mask,rtoi
-setMethod(f="cloud_mask",
-          signature = c("rtoi"),
-          function(x, products="ALL", verbose = FALSE, overwrite = FALSE,...){
-            if(products=="ALL") products<-product(x)
-            for(p in products){
-              img_dir<-get_mosaic_dir(x,p)
-              out_dir<-file.path(dirname(img_dir),"CloudMask")
-              dir.create(out_dir,showWarnings = FALSE)
-              all_files<-unlist(lapply(list.files(img_dir,full.names = TRUE,pattern="\\.zip$"), function(x){file.path("/vsizip",x,utils::unzip(x,list=TRUE)$Name)}))
+setMethod(
+  f = "cloud_mask",
+  signature = c("rtoi"),
+  function(x, products = "ALL", verbose = FALSE, overwrite = FALSE, ...) {
+    if (products == "ALL") products <- product(x)
+    for (p in products) {
+      img_dir <- get_mosaic_dir(x, p)
+      out_dir <- file.path(dirname(img_dir), "CloudMask")
+      dir.create(out_dir, showWarnings = FALSE)
+      all_files <- unlist(lapply(list.files(img_dir,
+                                            full.names = TRUE,
+                                            pattern = "\\.zip$"),
+                                 function(x) {
+        file.path("/vsizip", x, utils::unzip(x, list = TRUE)$Name)
+      }))
 
-              # Modis
-              if(any(grepl(tolower(substr(p,1,5)),c("mod09","myd09")))){
-                all_files<-all_files[grepl("_state_1km_1",all_files)]
-                fun_clkmsk<-modCloudMask
-              # Landsat
-              }else if(grepl("LANDSAT",p)){
-                all_files<-all_files[grepl("BQA",all_files)|grepl("pixel_qa",all_files)]
-                fun_clkmsk<-lsCloudMask
-                if(grepl("8",p)){
-                  cldcl <- c(322, 386, 834, 898, 1346, 324,  388,  836,  900, 1348)
-                }else{
-                  cldcl <- c(66, 130, 68, 132)
-                }
-              # Sentinel
-              }else if(p%in%unlist(SENPRODUCTS)){
-                all_files<-all_files[grepl("CLD",all_files)]
-                fun_clkmsk<-senCloudMask
-              }else{
-                if(verbose)message(paste0("Product", p," unsopported for cloud masking."))
-                next
-              }
+      # Modis
+      if (any(grepl(tolower(substr(p, 1, 5)), c("mod09", "myd09")))) {
+        all_files <- all_files[grepl("_state_1km_1", all_files)]
+        fun_clkmsk <- modCloudMask
+        # Landsat
+      } else if (grepl("LANDSAT", p)) {
+        all_files <- all_files[grepl("BQA", all_files) |
+                               grepl("pixel_qa", all_files)]
+        fun_clkmsk <- lsCloudMask
+        if (grepl("8", p)) {
+          cldcl <- c(322, 386, 834, 898, 1346, 324, 388, 836, 900, 1348)
+        } else {
+          cldcl <- c(66, 130, 68, 132)
+        }
+        # Sentinel
[…]
