+++ R/variables.R
@@ -1,104 +1,237 @@
-variables<-setRefClass("variables",
-                  # Define the slots
-                  fields = list(
-                    bands = "list"
-                  ),
-                  methods = list(
-                   EVI=function(blue,red,nir,scfun=function(r){r}){
-                      blue<-scfun(blue)
-                      red<-scfun(red)
-                      nir<-scfun(nir)
-                      evi <- 2.5*((nir - red) / (nir+6 * red-7.5*blue+1))
-                      return(evi)
-                    },
-                   MSAVI2=function(red, nir){
-                     msavi<-(2*nir+1-sqrt((2*nir+1)^2-8*(nir-red)))/2
-                     return(msavi)
-                   },
-                   NBR=function(nir,swir2){
-                     nbr<-(nir-swir2)/(nir+swir2)
-                     return(nbr)
-                   },
-                   NBR2=function(swir1,swir2){
-                     nbr2<-(swir1-swir2)/(swir1+swir2)
-                     return(nbr2)
-                   },
-                   NDMI=function(nir,swir1){
-                     ndmi<-(nir-swir1)/(nir+swir1)
-                     return(ndmi)
-                   },
-                   NDVI=function(red, nir){
-                     ndvi <- (nir - red) / (nir + red)
-                     return(ndvi)
-                   },
-                   NDWI=function(green,nir){
-                     ndwi <- ( green - nir ) / ( green + nir )
-                     return(ndwi)
-                   },
-                   RGB=function(red,green,blue,q.range=c()){
-                     rgb<-list(red,green,blue)
-                     names(rgb)<-c("red","green","blue")
+variables <- setRefClass("variables",
+  # Define the slots
+  fields = list(
+    bands = "list"
+  ),
+  methods = list(
+    EVI = function(blue, red, nir, scfun = function(r) {
+                     r
+                   }) {
+      blue <- scfun(blue)
+      red <- scfun(red)
+      nir <- scfun(nir)
+      evi <- 2.5 * ((nir - red) / (nir + 6 * red - 7.5 * blue + 1))
+      return(evi)
+    },
+    MSAVI2 = function(red, nir) {
+      msavi <- (2 * nir + 1 - sqrt((2 * nir + 1)^2 - 8 * (nir - red))) / 2
+      return(msavi)
+    },
+    NBR = function(nir, swir2) {
+      nbr <- (nir - swir2) / (nir + swir2)
+      return(nbr)
+    },
+    NBR2 = function(swir1, swir2) {
+      nbr2 <- (swir1 - swir2) / (swir1 + swir2)
+      return(nbr2)
+    },
+    NDMI = function(nir, swir1) {
+      ndmi <- (nir - swir1) / (nir + swir1)
+      return(ndmi)
+    },
+    NDVI = function(red, nir) {
+      ndvi <- (nir - red) / (nir + red)
+      return(ndvi)
+    },
+    NDWI = function(green, nir) {
+      ndwi <- (green - nir) / (green + nir)
+      return(ndwi)
+    },
+    RGB = function(red, green, blue, q.range = c()) {
+      rgb <- list(red, green, blue)
+      names(rgb) <- c("red", "green", "blue")
 
-                     if(!is.null(q.range)){
-                       rgb<-lapply(rgb,FUN = function(r,q.range){q<-raster::quantile(r,q.range,na.rm=TRUE);r<-clamp(r,lower=q[1],upper=q[2]);return(r)},q.range)
-                       names(rgb)<-c("red","green","blue")
-                     }
-                     rgb<-lapply(rgb,raster::stretch, minv=0, maxv=255)
-                     image<-raster::stack(rgb)
[…]
