+++ R/add2rtoi.R
@@ -5,19 +5,22 @@ setGeneric("add2rtoi", function(infile,
   standardGeneric("add2rtoi")
 })
 
-setMethod(f="add2rtoi",
-          signature = c("character","character"),
-          function(infile, out.zip, ...){
-            if(!file.exists(infile)){
-              warning(paste0("File not created: ",infile))
-              return(NULL)
-            }
-            if(!file.exists(out.zip)){
-              zip::zipr(out.zip,files=infile)
-            }else{
-              zip::zipr_append(out.zip, infile, recurse = TRUE,
-                               include_directories = TRUE)
-            }
-            file.remove(infile)
-})
-
+setMethod(
+  f = "add2rtoi",
+  signature = c("character", "character"),
+  function(infile, out.zip, ...) {
+    if (!file.exists(infile)) {
+      warning(paste0("File not created: ", infile))
+      return(NULL)
+    }
+    if (!file.exists(out.zip)) {
+      zip::zipr(out.zip, files = infile)
+    } else {
+      zip::zipr_append(out.zip, infile,
+        recurse = TRUE,
+        include_directories = TRUE
+      )
+    }
+    file.remove(infile)
+  }
+)

[…]
