### 2021-04-09 7ee653e
user seq_len instead 1:n
