### 2021-04-09 #437 @jhollist
Sounds good.  Let me get some other opinions on the pre-review CRAN
submission.  Stay tuned.

On Fri, Apr 9, 2021 at 1:05 PM Unai Pérez-Goya ***@***.***>
wrote:

> @jhollist <https://github.com/jhollist> thanks for your comments.
>
> I think we can deal with most of the changes in a short period of time. I
> think the biggest change is the testing. We will try to get to the 70% that
> you mention.
>
> I was wondering if I could send the current version of the package to
> CRAN. It would be great for us to report that for a project. However, we
> are concerned about the change of the name of the package during the
> ropensci revision. Therefore, we would like to receive your advice on this
> regard.
>
