+++ R/data.R
@@ -2,8 +2,7 @@
 #'
 #' Geographically projected \code{RasterStack} with the digital elevation model
 #' (DEM) of the region of Navarre (Spain). The DEM was obtained from the
-#' \href{http://centrodedescargas.cnig.es/CentroDescargas/
-#' locale?request_locale=en#}{National Center for Geographic Information}
+#' \href{http://centrodedescargas.cnig.es/CentroDescargas/locale?request_locale=en}{National Center for Geographic Information}
 #' of Spain. The DEM is used as a covariable in the Image Mean Anomaly (IMA)
 #' algorithm (\code{\link{smoothing_images}}).
 #'
