+++ tests/testthat/test-credentials.R
@@ -1,7 +1,0 @@
-test_that("credentials test", {
-  set_credentials("username", "password")
-
-  set_credentials("username","password","scihub")
-
-  print_credentials()
-})
