+++ tests/testthat/test-rtoi.R
@@ -1,16 +1,6 @@
-test_that("api test", {
+test_that("records test", {
   data(ex.navarre)
-
-  set_credentials("username","password")
-
-  records<-sat_search(
-    region = ex.navarre,
-    product = c("mod09ga", "LANDSAT_8_C1","S2MSI2A"),
-    dates = as.Date("2021-03-01") + seq(1, 20),
-    verbose=TRUE,
-    test.mode=TRUE
-  )
-
+  set_credentials("username", "password")
   rtoi.path <- tempdir()
 
   show_variables()
