+++ man/sat_search.Rd
@@ -10,9 +10,9 @@
 \usage{
 sat_search(region, product, ...)
 
-\S4method{sat_search}{rtoi,character}(region, product, verbose = FALSE, ...)
+\S4method{sat_search}{rtoi,character}(region, product, verbose = FALSE, test.mode = FALSE, ...)
 
-\S4method{sat_search}{sf,character}(region, product, verbose = FALSE, ...)
+\S4method{sat_search}{sf,character}(region, product, verbose = FALSE, test.mode = FALSE, ...)
 }
 \arguments{
 \item{region}{a \code{Spatial*}, \code{Raster*}, \code{sf} or \code{rtoi}
@@ -24,6 +24,8 @@ class objects defining the region of interest.}
 
 \item{verbose}{logical argument. If \code{TRUE}, the function prints the
 running steps and warnings.}
+
+\item{test.mode}{logical argument. If \code{TRUE}, the function gets test data from github.}
 }
 \description{
 Search satellite images concerning a particular location, data product, and
