+++ R/preview.R
@@ -1,28 +1,29 @@
-#' @import  leaflet leafem
+#' @importFrom leaflet addProviderTiles addLayersControl addPolygons
+#' @importFrom leaflet providerTileOptions layersControlOptions
 createMap <- function() {
-  leaflet() %>%
-    addProviderTiles("OpenStreetMap",
-                     group = "OpenStreetMap",
-                     options = providerTileOptions(zIndex = 0)) %>%
-    addProviderTiles("Esri.WorldImagery",
-                     group = "Esri.WorldImagery",
-                     options = providerTileOptions(zIndex = 0)) %>%
-    addProviderTiles("CartoDB.Positron",
-                     group = "CartoDB.Positron",
-                     options = providerTileOptions(zIndex = 0)) %>%
-    addProviderTiles("CartoDB.DarkMatter",
-                     group = "CartoDB.DarkMatter",
-                     options = providerTileOptions(zIndex = 0)) %>%
-    addLayersControl(
-      baseGroups = c(
-        "OpenStreetMap",
-        "Esri.WorldImagery",
-        "CartoDB.Positron",
-        "CartoDB.DarkMatter"
-      ),
-      position = "topleft",
-      options = layersControlOptions(autoZIndex = FALSE)
-    )
+  theMap<-leaflet::leaflet()
+  theMap<-addProviderTiles(theMap,"OpenStreetMap",
+                           group = "OpenStreetMap",
+                           options = providerTileOptions(zIndex = 0))
+  theMap<-addProviderTiles(theMap,"Esri.WorldImagery",
+                           group = "Esri.WorldImagery",
+                           options = providerTileOptions(zIndex = 0))
+  theMap<-addProviderTiles(theMap,"CartoDB.Positron",
+                           group = "CartoDB.Positron",
+                           options = providerTileOptions(zIndex = 0))
+  theMap<-addProviderTiles(theMap,"CartoDB.DarkMatter",
+                           group = "CartoDB.DarkMatter",
+                           options = providerTileOptions(zIndex = 0))
+  addLayersControl(theMap,
+    baseGroups = c(
+      "OpenStreetMap",
+      "Esri.WorldImagery",
+      "CartoDB.Positron",
+      "CartoDB.DarkMatter"
+    ),
+    position = "topleft",
+    options = layersControlOptions(autoZIndex = FALSE)
+  )
 }
 
 updateLayersControl <- function(map,
@@ -39,10 +40,13 @@ updateLayersControl <- function(map,
   return(map)
 }
 
+#' @importFrom leafem updateLayersControl addRasterRGB
+#' @importFrom stars st_as_stars
 addMapRasterRGB <- function(img, lname, lpos, project = project) {
-  nmap <- getPreviewMap() %>%
+  nmap <- getPreviewMap() #%>%
     # addRasterRGB(subset(img,lpos), group=lname) #%>%
-    addRasterRGB(st_as_stars(subset(img, lpos)),
+  nmap <- addRasterRGB(nmap,
+                 st_as_stars(subset(img, lpos)),
                  group = lname,
                  project = project)
   # addHomeButton(st_bbox(transform_multiple_proj(img,
@@ -54,6 +58,9 @@ addMapRasterRGB <- function(img, lname, lpos, project = project) {
   setPreviewMap(updateLayersControl(nmap, addOverlayGroups = lname))
 }
 
+#' @importFrom leaflet addPolygons
+#' @importFrom leafem addHomeButton addFeatures
+#' @importFrom sf st_bbox st_crs
 addMapFeature <- function(sfobj, lname) {
   nmap <- getPreviewMap()
[…]
