+++ NAMESPACE
@@ -48,10 +48,7 @@ exportMethods(subset)
 exportMethods(unique)
 import(Rdpack)
 import(XML)
-import(curl)
 import(fields)
-import(leafem)
-import(leaflet)
 import(methods)
 import(raster)
 import(rjson)
@@ -60,15 +57,37 @@ import(stars)
 import(tmap)
 import(zip)
 importFrom(calendR,calendR)
+importFrom(curl,curl)
+importFrom(curl,curl_download)
+importFrom(curl,curl_fetch_memory)
+importFrom(curl,handle_setform)
+importFrom(curl,handle_setopt)
+importFrom(curl,new_handle)
 importFrom(grDevices,colors)
 importFrom(httr,POST)
 importFrom(httr,add_headers)
 importFrom(httr,authenticate)
 importFrom(httr,content)
 importFrom(httr,content_type)
+importFrom(leafem,addFeatures)
+importFrom(leafem,addHomeButton)
+importFrom(leafem,addRasterRGB)
+importFrom(leafem,updateLayersControl)
+importFrom(leaflet,addLayersControl)
+importFrom(leaflet,addPolygons)
+importFrom(leaflet,addProviderTiles)
+importFrom(leaflet,layersControlOptions)
+importFrom(leaflet,providerTileOptions)
+importFrom(raster,extent)
+importFrom(raster,stack)
+importFrom(raster,writeRaster)
 importFrom(rvest,html_nodes)
+importFrom(sf,gdal_utils)
+importFrom(sf,st_bbox)
+importFrom(sf,st_crs)
 importFrom(sp,"proj4string<-")
 importFrom(sp,proj4string)
+importFrom(stars,st_as_stars)
 importFrom(tools,md5sum)
 importFrom(utils,URLencode)
 importFrom(utils,untar)
