+++ tests/testthat/test-rtoi.R
@@ -8,14 +8,20 @@ test_that("records test", {
 
   # path where downloads are stored
   db.path <- file.path(tempdir(), "DATABASE")
+  unlink(file.path(rtoi.path,"Navarre"),recursive = T)
+
+
   navarre <- new_rtoi(
-    "Navarre",
+    "Navarre_rtoi",
     ex.navarre,
     rtoi.path,
     db.path
   )
 
+
   navarre <- read_rtoi(file.path(rtoi.path,"Navarre"))
+
+
 
   #######################################################
   # Searching errors
@@ -90,5 +96,5 @@ test_that("records test", {
 
   preview(navarre,dates(navarre)[2])
 
-  unlink(file.path(rtoi.path,"Navarre"),recursive = T)
+  unlink(file.path(rtoi.path,"Navarre_rtoi"),recursive = T)
 })
