+++ tests/testthat/test-download.R
@@ -29,6 +29,7 @@ test_that("download test", {
   plot(navarre, "view", product = unique(product(navarre))[2])
   plot(navarre, "view", product = unique(product(navarre))[3])
 
+  print(navarre)
   list_data(navarre)
 
   derive(navarre,product="LANDSAT_8_C1_lvl2",variable="NDVI")
@@ -36,6 +37,7 @@ test_that("download test", {
   derive(navarre,product="S2MSI2A",variable="NDVI")
   plot(navarre,"view",variable="NDVI",product = unique(product(navarre))[1])
   plot(navarre,"view",variable="NDVI",product = unique(product(navarre))[2])
+
   #plot(navarre,"view",variable="NDVI",product = unique(product(navarre))[3])# derive with s2
   navarre
   list_data(navarre)
