+++ R/list_data.R
@@ -84,6 +84,7 @@ setMethod("list_data",
         if (any(grepl("mosaic", allproducts))) {
           f <- list.files(allproducts[grepl("mosaic", allproducts)],
                           full.names = TRUE)[1]
+          if(is.na(f)) return(NULL)
           vars <- gsub("\\s*(\\d{7}_)", "", utils::unzip(f, list = TRUE)$Name)
           vars <- gsub("\\.tif$", "", vars)
           f <- gsub(paste0(get_dir(x), "/"), "", f)
