+++ tests/testthat/test-smoothing_IMA.R
@@ -3,9 +3,6 @@ test_that("smoothing IMA", {
   # load an example of NDVI time series in Navarre
   data(ex.ndvi.navarre)
 
-  # the raster stack with the date in julian format as name
-  spplot(ex.ndvi.navarre)
-
   # smoothin and fill all the time series
 
   tiles.mod.ndvi.filled <- smoothing_images(ex.ndvi.navarre,
