+++ R/plot.R
@@ -544,31 +544,8 @@ genPlotGIS <- function(r,
                 "RasterStack, or a list of RasterBrick or RasterStack."))
   }
 
-
-
   # layout preconfigured arguments
-  tm_layout_args <- args[names(args) %in% names(formals(tm_layout))]
-  if (!("legend.bg.color" %in% names(tm_layout_args))) {
-    tm_layout_args$legend.bg.color <- "white"
-  }
-  if (!("panel.show" %in% names(tm_layout_args))) {
-    tm_layout_args$panel.show <- TRUE
-  }
-  if (!("panel.labels" %in% names(tm_layout_args))) {
-    tm_layout_args$panel.labels <- names(r)
-  }
-  if (!("legend.outside" %in% names(tm_layout_args))) {
-    tm_layout_args$legend.outside <- TRUE
-  }
-  if (!("legend.outside.size" %in% names(tm_layout_args))) {
-    tm_layout_args$legend.outside.size <- 0.08
-  }
-  if (!("legend.outside.position" %in% names(tm_layout_args))) {
-    tm_layout_args$legend.outside.position <- "right"
-  }
-  if (!("frame" %in% names(tm_layout_args))) {
-    tm_layout_args$frame <- TRUE
-  }
+  tm_layout_args<-initialize.tm.layout(names(r),...)
 
   if (!missing(layout)) {
     lyt <- tm_facets(ncol = layout[2], nrow = layout[1])
@@ -576,43 +553,13 @@ genPlotGIS <- function(r,
     lyt <- NULL
   }
 
-  graticules_args <- c(names(formals(tm_graticules)), names(formals(tm_grid)))
-  names(graticules_args) <- paste0("tm.graticules.", graticules_args)
-  tm_graticules_args <- args[names(args) %in% names(graticules_args)]
-  names(tm_graticules_args) <- graticules_args[names(tm_graticules_args)]
-  if (!("lines" %in% names(tm_graticules_args))) {
-    tm_graticules_args$lines <- FALSE
-  }
-  if (as.grid) {
-    tm_layout_args$between.margin <- -.1
-    if (!("labels.space.x" %in% names(tm_layout_args))) {
-      tm_graticules_args$labels.space.x <- .10
-    }
-    if (!("labels.space.y" %in% names(tm_layout_args))) {
-      tm_graticules_args$labels.space.y <- .10
-    }
-  }
-  grid <- do.call(tm_graticules, tm_graticules_args)
+  # create grid
+  grid <- create.tm.grid(as.grid,...)
+
 
   # compass arguments and preconfigured assignation
   if (!compass.rm) {
-    compass_args <- names(formals(tm_compass))
-    names(compass_args) <- paste0("tm.compass.", compass_args)
-    tm_compass_args <- args[names(args) %in% names(compass_args)]
-    names(tm_compass_args) <- compass_args[names(tm_compass_args)]
-    if (!("type" %in% names(tm_compass_args))) {
-      tm_compass_args$type <- "arrow"
-    }
-    if (!("position" %in% names(tm_compass_args))) {
-      tm_compass_args$position <- c("right", "top")
-    }
-    if (!("size" %in% names(tm_compass_args))) {
-      tm_compass_args$size <- 2
-    }
-    if (!("show.labels" %in% names(tm_compass_args))) {
-      tm_compass_args$show.labels <- 0
-    }
-    compass <- do.call(tm_compass, tm_compass_args)
+    compass <- create.compass(...)
   } else {
     compass <- NULL
   }
@@ -620,52 +567,13 @@ genPlotGIS <- function(r,
 
   # scale bar arguments and preconfigured assignation
   if (!scale.bar.rm) {
[…]
