+++ R/mosaic_fun_SYN.R
@@ -1,163 +1,163 @@
-mosaic_fun_SY_2_SYN <- function(mfiles, ...) {
-  # TODO
-  # "SY_2_SYN___"
-  # if(product(dr[1])="OL_2_LFR___"){
-  #   readfromscratch<-function(m){
-  #     # extract
-  #     # get new paths
-  #     # filtra las bandas que queramos
-  #     # return the paths
-  #   }
-  #   filterchunks<-function(allfiles,bnds){
-  #     allfiles[grepl(paste0(bnds,".jp2"),allfiles, ignore.case=TRUE)]
-  #   }
-  #   bands<-c("ogvi.nc","tie_meteo.nc")#manual designation of bands
-  #
-  # }
-
-  allfun <- NULL
-  allfun$readfromscratch <- function(m, ...) {
-    args <- list(...)
-    if ("bands.files" %in% names(args)) {
-      bands.files <- args$bands.files
-    } else {
-      stop("Error reading from scrach.")
-    }
-    ext.dir <- file.path(args$scratch.tmp, gsub(".zip", "", basename(m)))
-    unzip(m, exdir = ext.dir, junkpaths = TRUE)
-    # newpaths<-list.files(ext.dir,full.names = TRUE)
-    newpaths <- file.path(ext.dir, paste0(bands.files, ".nc"))
-
-    # extract coordinates
-    coordinates.file <- file.path(ext.dir, "geolocation.nc")
-    coor.info <- gdal_utils(
-      util = "info",
-      source = coordinates.file,
-      quiet = TRUE
-    )
-
-    coor.info <- unlist(strsplit(coor.info, "\n"))
-    coor.names <- coor.info[grepl(".*SUBDATASET_.*_NAME=", coor.info)]
-    coor.names <- gsub(".*SUBDATASET_.*_NAME=", "", coor.names)
-    coor.names <- gsub("\"", "", coor.names, fixed = TRUE)
-
-    lat <- file.path(ext.dir, "lat.vrt")
-    lon <- file.path(ext.dir, "lon.vrt")
-    gdal_utils(
-      util = "translate",
-      source = coor.names[2],
-      destination = lat,
-      options = c("-of", "VRT")
-    )
-    gdal_utils(
-      util = "translate",
-      source = coor.names[3],
-      destination = lon,
-      options = c("-of", "VRT")
-    )
-
-    # TODO remove GCP from vrt
-    # removeGCPfromVRT<-function(fpath){
-    #  a<-read_xml(fpath)
-    # }
-
-    newbands <- c()
-    for (np in newpaths) {
-      image.data <- gdal_utils(
-        util = "info",
-        source = np, quiet = TRUE
-      )
-      image.data <- unlist(strsplit(image.data, "\n"))
-      bands.names <- image.data[grepl(".*SUBDATASET_.*_NAME=", image.data)]
-      bands.names <- gsub(".*SUBDATASET_.*_NAME=", "", bands.names)
-      bands.names <- gsub("\"", "", bands.names, fixed = TRUE)
-      img.name <- gsub("\\.nc", "", basename(np))
-
-      # add projection
-      for (b in bands.names) {
-        b.tmp.name <- gsub(":", "_", paste0(gsub("\\.", "_", basename(b)),
-                                            ".vrt"))
-        b.name <- gsub("\\.vrt", ".tif", b.tmp.name)
-        out.tmp.file.name <- file.path(ext.dir, b.tmp.name)
-        print(out.tmp.file.name)
-        out.file.name <- file.path(ext.dir, b.name)
-        if (!file.exists(out.file.name)) {
-          gdal_utils(
-            util = "translate",
[…]
