### 2021-04-20 74125ed
remove SY_2_SYN___ mosaic
