+++ man/test_function.Rd
@@ -1,0 +1,11 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/testing_function.R
+\name{test_function}
+\alias{test_function}
+\title{Testing function}
+\usage{
+test_function()
+}
+\description{
+Function used for testing some internal functions in continious integration.
+}
