+++ NAMESPACE
@@ -32,6 +32,7 @@ export(set_credentials)
 export(set_database)
 export(show_variables)
 export(smoothing_images)
+export(test_function)
 exportClasses(rtoi)
 exportMethods("[")
 exportMethods(as.data.frame)
