+++ .github/workflows/R-CMD-check.yaml
@@ -23,7 +23,7 @@ jobs:
       matrix:
         config:
           - {os: windows-latest, r: 'release'}
-          #- {os: macOS-latest, r: 'release'} # qpdf deploy error
+          - {os: macOS-latest, r: 'release'} # qpdf deploy error
           - {os: ubuntu-20.04, r: 'release', rspm: "https://packagemanager.rstudio.com/cran/__linux__/focal/latest"}
           - {os: ubuntu-20.04, r: 'devel', rspm: "https://packagemanager.rstudio.com/cran/__linux__/focal/latest"}
 
