+++ tests/testthat/test-smoothing_IMA.R
@@ -8,14 +8,12 @@ test_that("smoothing IMA", {
   tiles.mod.ndvi.filled <- smoothing_images(ex.ndvi.navarre,
                                             method = "IMA",
                                             only.na = TRUE,
-                                            predictSE = FALSE,
-                                            snow.mode = FALSE
+                                            predictSE = FALSE
   )
   tiles.mod.ndvi.filled <- smoothing_images(ex.ndvi.navarre,
                                             method = "IMA",
                                             only.na = TRUE,
-                                            predictSE = TRUE,
-                                            snow.mode = TRUE
+                                            predictSE = TRUE
   )
 
 
