### 2021-04-22 6f7db5e
import clamp from raster
