+++ NAMESPACE
@@ -90,6 +90,7 @@ importFrom(raster,addLayer)
 importFrom(raster,aggregate)
 importFrom(raster,as.matrix)
 importFrom(raster,beginCluster)
+importFrom(raster,clamp)
 importFrom(raster,clusterR)
 importFrom(raster,crs)
 importFrom(raster,endCluster)
