+++ man/ex.dem.navarre.Rd
@@ -20,7 +20,7 @@ The \code{RasterStack} coordinates are in the Sinusoidal projection.
 Geographically projected \code{RasterStack} with the digital elevation model
 (DEM) of the region of Navarre (Spain). The DEM was obtained from the
 \href{http://centrodedescargas.cnig.es/CentroDescargas/locale?request_locale=en}{National Center for Geographic Information}
-of Spain. The DEM is used as a covariable in the Image Mean Anomaly (IMA)
+of Spain. The DEM is used as a covariate in the Image Mean Anomaly (IMA)
 algorithm (\code{\link{smoothing_images}}).
 }
 \keyword{data}
