+++ NAMESPACE
@@ -142,6 +142,7 @@ importFrom(tmap,tm_shape)
 importFrom(tmap,tmap_arrange)
 importFrom(tools,md5sum)
 importFrom(utils,URLencode)
+importFrom(utils,data)
 importFrom(utils,head)
 importFrom(utils,untar)
 importFrom(utils,unzip)
