+++ NAMESPACE
@@ -104,6 +104,7 @@ importFrom(raster,nlayers)
 importFrom(raster,predict)
 importFrom(raster,projectRaster)
 importFrom(raster,projection)
+importFrom(raster,quantile)
 importFrom(raster,raster)
 importFrom(raster,readAll)
 importFrom(raster,stack)
