### 2021-04-28 #437 @lolaugartemartinez
Gracias Unai!

El Mar, 27 de Abril de 2021, 17:42, Unai Pérez-Goya escribió:
> @jhollist thanks for your comments. We decided not to send the package to
> CRAN until the review process is completed.
> We covered most of the requirements. Find below the line by line answers.
>
> ##### Automated Tests
>
>     * [x]  `*`Expand CI to include Windows, Linux, and Mac and using the
> latest, previous and development versions of R.  You are currently
> using appveyor which is using the latest version and Windows only.
> You may want to look at GitHub Actions which is supported via
> `usethis::use_github_action_check_standard()`.  This will test on all
> the needed platforms and releases.
>
> Done, we integrated the package in GitHub Actions. Additionally, we change
> the minimum R version supported by the package to 4.0.0. Packages we
> depend on have experienced major changes since the 4.0.0. R version, which
> means that rsat is no longer backward compatible for earlier versions.
>
> ##### Masked function names
>
>     * [ ]  I did notice a few masked function names (e.g. `rsat::subset`).
>  I agree that masking is sometimes hard to avoid, but it might be
> worthwhile to rename is possible.  It is likely that reviewers might
> catch this as well.
>
> These masked functions are developed to add functionalities to the new
> classes created by the package. For example, rsat::subset gets a subset
> from `records`, that works as vector.
> We created new methods for these functions, so they accept rtois or
> records as inputs. When used with other classes, their original
> functionalities remain the same. We think that this strategy benefits R
> users as they do not need to remember new names specific for
> rtois/records.
>
>
> ##### Checks and Tests
>
> Tests (e.g. `devtools::check()` and `devtools::test()`) are currently
> failing on my machine, plus I am seeing several warnings.
>
>     * [x]  `*`Check warnings on tests.  Some might be acceptable but
> others may need
>       correction.  I see:
>
> Done. Now only a few warnings related to modis image projections are
> displayed
>
> ```
> -- Warning (test-search.R:17:5): search test
> -----------------------------------
> Error in Earth Explorer api connection. HTTP 404.
> Backtrace:
>  1. rsat::sat_search(...) test-search.R:17:4
>  2. rsat::sat_search(...)
>  3. rsat:::.local(region, product, ...)
>  5. rsat:::ls_search(region(region), product = s, ...)
>  6. rsat:::.local(region, product, ...)
> -- Warning (test-search.R:17:5): search test
> -----------------------------------
> Error in Earth Explorer api connection. HTTP 404.
> Backtrace:
>  1. rsat::sat_search(...) test-search.R:17:4
>  2. rsat::sat_search(...)
>  3. rsat:::.local(region, product, ...)
>  5. rsat:::ls_search(region(region), product = s, ...)
>  6. rsat:::.local(region, product, ...)
> -- Warning (test-search.R:17:5): search test
> -----------------------------------
> Error in Earth Explorer api connection. HTTP 404.
> Backtrace:
>  1. rsat::sat_search(...) test-search.R:17:4
>  2. rsat::sat_search(...)
>  3. rsat:::.local(region, product, ...)
>  5. rsat:::ls_search(region(region), product = s, ...)
>  6. rsat:::.local(region, product, ...)
> -- Warning (test-search.R:17:5): search test
> -----------------------------------
> Error in Earth Explorer api connection. HTTP 404.
> Backtrace:
>  1. rsat::sat_search(...) test-search.R:17:4
>  2. rsat::sat_search(...)
>  3. rsat:::.local(region, product, ...)
>  5. rsat:::ls_search(region(region), product = s, ...)
>  6. rsat:::.local(region, product, ...)
> -- Warning (test-search.R:17:5): search test
> -----------------------------------
> Error in Earth Explorer api connection. HTTP 404.
> Backtrace:
>  1. rsat::sat_search(...) test-search.R:17:4
>  2. rsat::sat_search(...)
>  3. rsat:::.local(region, product, ...)
>  5. rsat:::ls_search(region(region), product = s, ...)
>  6. rsat:::.local(region, product, ...)
> -- Warning (test-search.R:17:5): search test
> -----------------------------------
> Error in Earth Explorer api connection. HTTP 404.
> Backtrace:
>  1. rsat::sat_search(...) test-search.R:17:4
>  2. rsat::sat_search(...)
>  3. rsat:::.local(region, product, ...)
>  5. rsat:::ls_search(region(region), product = s, ...)
>  6. rsat:::.local(region, product, ...)
> -- Warning (test-search.R:17:5): search test
> -----------------------------------
> Cannot perform Landsat search, check your credentials and/or the api
> status: https://m2m.cr.usgs.gov/api/docs/json/
> Backtrace:
>  1. rsat::sat_search(...) test-search.R:17:4
>  2. rsat::sat_search(...)
>  3. rsat:::.local(region, product, ...)
>  5. rsat:::ls_search(region(region), product = s, ...)
>  6. rsat:::.local(region, product, ...)
> -- Warning (test-search.R:35:5): search test
> -----------------------------------
> GDAL Error 1: PROJ: proj_create_from_database: crs not found
> Backtrace:
>  1. base::plot(c(rcds1, rcds2, rcds3)) test-search.R:35:4
>  2. rsat::plot(c(rcds1, rcds2, rcds3))
>  3. rsat:::.local(x, y, ...)
>  5. sf:::st_crs.numeric(crs(r))
>  6. sf:::make_crs(paste0("EPSG:", x))
>  7. sf:::CPL_crs_from_input(x)
> ```
>
>     * [x]  `*`Make sure tests pass.  I see:
>
> Done.
>
> ```
> == Failed tests
> ================================================================
> -- Error (test-search.R:35:5): search test
> -------------------------------------
> Error: attempt to select less than one element in get1index
> Backtrace:
>     x
>  1. +-base::plot(c(rcds1, rcds2, rcds3)) test-search.R:35:4
>  2. \-rsat::plot(c(rcds1, rcds2, rcds3))
>  3.   \-rsat:::.local(x, y, ...)
>  4.     \-connection$getApi(get_api_name(r))
> ```
>
> ##### Goodpractice
>
> Many issues are raised with `goodpractice::gp()`. Pay close attention to
> the output and fix these issues. In particular:
>
>     * [x]  Consider simplifying `genPlotGIS`
> Done. The function was divided into smaller functions.
>
> ```
>   ✖ write short and simple functions. These functions have high
>     cyclomatic complexity:genPlotGIS (60).
> ```
>
>
>
>
>     * [x]  `*`You currently have sf, raster, and stars as Depends.  Those
> can all be moved to Imports.  You will need to make sure to import
> those (e.g. with @importFrom) where you use them in your functions.
> Done. All the packages as Depends were passed to imports. Now the package
> only uses importFrom.
>
>
> ```
>   ✖ not use "Depends" in DESCRIPTION, as it can cause name
>     clashes, and poor interaction with other packages. Use "Imports"
>     instead.
> ```
>     * [x]  `*`Add a URL field.  Prior to acceptance as an ROpenSci package
> this would be https://github.com/spatialstatisticsupna/rsat.
> Done.
>
>
> ```
>   ✖ add a "URL" field to DESCRIPTION. It helps users find
>     information about your package online. If your package does not have
> a
>     homepage, add an URL to GitHub, or the CRAN package package page.
> ```
>     * [x]  `*`Add BugReports field.  Prior to acceptance as an ROpenSci
> package this would be
> https://github.com/spatialstatisticsupna/rsat/issues.
> Done.
>
> ```
>   ✖ add a "BugReports" field to DESCRIPTION, and point it to a bug
>     tracker. Many online code hosting services provide bug trackers for
>     free, https://github.com, https://gitlab.com, etc.
> ```
>
>
>
>     * [x]  `*`It appears you are using both '=' and '<-' for assignment.
> The ROpenSci style guide allows either, but you need to be consistent.
>  Please choose one and use it throughout your code.
>
> Done.
> ```
>   ✖ use '<-' for assignment instead of '='. '<-' is the standard,
>     and R users and developers are used it and it is easier to read your
>     code for them if you use '<-'.
>
>     R\api.R:21:21
>     R\api.R:22:21
>     R\api.R:23:20
>     R\api.R:24:20
>     R\api.R:25:24
>     ... and 102 more lines
> ```
>
>     * [x]  Work on getting your long lines shortened.  The `styler`
> package will help with some of these, but not all (e.g.
> [r-lib/styler#247](https://github.com/r-lib/styler/issues/247)).  You
> will have to manually edit many of these.  You can use
> `lintr::lint(path/to/file.R)` to help find these.  Also, you can turn
> on an 80 character guide in RStudio with Tools:Global
> Options:Code:Display and select Show Margin and set characters to 80.
> This will help visually identify these lines.
>
> Done. There are some lines with more than 80 characters, but these are
> URLs that it is not possible to cut.
>
> ```
>   ✖ avoid long code lines, it is bad for readability. Also, many
>     people prefer editor windows that are about 80 characters wide. Try
> make
>     your lines shorter than 80 characters
>
>     R\api.R:6:1
>     R\api.R:118:1
>     R\api.R:120:1
>     R\api.R:133:1
>     R\api.R:150:1
>     ... and 321 more lines
> ```
>
>     * [x]  `*`Omit any trailing semicolons.
>
> Done. We fixed this.
>
> ```
>   ✖ omit trailing semicolons from code lines. They are not needed
>     and most R coding standards forbid them
>
>     R\mosaic.R:166:92
>     R\preview.R:217:40
> ```
>
>     * [x]  `*`Replace `sapply()` with `vapply()`
>
> Done. Now the package uses vapply.
> ```
>   ✖ avoid sapply(), it is not type safe. It might return a vector,
>     or a list, depending on the input data. Consider using vapply()
> instead.
>
>     R\package_tools.R:40:10
>     R\search_ls.R:107:17
> ```
>
>     * [x]  `*`When building a sequence use `seq_len()` or `seq_along()`
> Done. Now the package uses seq_len or seq_along.
>
> ```
>   ✖ avoid 1:length(...), 1:nrow(...), 1:ncol(...), 1:NROW(...) and
>     1:NCOL(...) expressions. They are error prone and result 1:0 if the
>     expression on the right hand side is zero. Use seq_len() or
> seq_along()
>     instead.
>
>     R\download.R:74:22
>     R\download.R:107:24
>     R\download.R:125:24
>     R\mosaic_generic.R:61:103
>     R\package_tools.R:42:12
>     ... and 9 more lines
> ```
>
>     * [x]  `*`Refine importing of functions.  Instead of @import, use
> @importFrom package function_name
> Done, now we only use importFrom.
>
> ```
>   ✖ do not import packages as a whole, as this can cause name clashes
>     between the imported packages. Instead, import only the specific
>     functions you need.
> ```
>
>     * [x]  `*`Same testing error as mentioned above.
> Fixed.
>
> ```
>   ✖ checking tests ... Running 'testthat.R' ERROR Running the
>     tests in 'tests/testthat.R' failed. Last 13 lines of output: Data and
>     variable methods provided by rsat Satellite products: ls1, ls2, ls3,
>     ls4, ls5, ls7, ls8, mod09ga, myd09ga, mcd43a4, Sentinel-1,
> Sentinel-2,
>     Sentinel-3, SY_2_SYN___. Variable Methods: EVI, MSAVI2, NBR, NBR2,
> NDMI,
>     NDVI, NDWI, RGB, SAVI.== Failed tests
>     ================================================================ --
>     Error (test-search.R:35:5): search test
>     ------------------------------------- Error: attempt to select less
> than
>     one element in get1index Backtrace: x 1. +-base::plot(c(rcds1, rcds2,
>     rcds3)) test-search.R:35:4 2. \-rsat::plot(c(rcds1, rcds2, rcds3)) 3.
>     \-rsat:::.local(x, y, ...) 4.  \-connection$getApi(get_api_name(r)) [
>     FAIL 1 | WARN 8 | SKIP 0 | PASS 0 ] Error: Test failures Execution
>     halted
> ```
>
>     * [x]  I only see this one with `goodpractice::gp()`, not when I run
> check on its own.  You can probably ignore this one.
>
> Running  goodpractice::gc(), there is no warning anymore.
> ```
>   ✖ fix this R CMD check WARNING: LaTeX errors when creating PDF
>     version. This typically indicates Rd problems. LaTeX errors found:
> ```
>
>     * [x]  `*`Switch T and F to TRUE and FALSE
> Fixed! We changed all T and F as TRUE and FALSE.
>
> ```
>   ✖ avoid 'T' and 'F', as they are just variables which are set to
>     the logicals 'TRUE' and 'FALSE' by default, but are not reserved
> words
>     and hence can be overwritten by the user.  Hence, one should always
> use
>     'TRUE' and 'FALSE' for the logicals.
>
>     R/api.R:NA:NA
>     R/cloud_mask.R:NA:NA
>     R/cloud_mask.R:NA:NA
>     R/derive.R:NA:NA
>     R/derive.R:NA:NA
>     ... and 19 more lines
> ```
>
> ##### Spelling
>
>     * [x]  Several items are flagged with `devtools::spell_check()` but
> most appear to be names, acronyms, etc. and are probably not really
> misspellings.  Please double check the items returned by
> `devtools::spell_check()` to make sure any true misspellings are
> corrected.
>
> We check all the misspellings
>
> ##### Tests
>
>     * [x]  `*`Expand tests: Since I had the one test fail locally, I
> wasn't able to confirm coverage.  On your README, you report 54% and
> it appears that many of the exported functions in your package are not
> covered by your current tests.  I would like to see the testing
> expanded to 70% or greater.  I do realize that testing with API
> packages can be a challenge sometimes, so if reaching 70% is a problem
> you will need to provide a justification.  Additionally, if
> authentication or API keys are required you need to include
> instructions in your documentation (e.g. in README).  In short, if
> there are any prerequisites for your tests to pass, makes sure users
> and reviewers can easily set those up on their local machines.
>
> Testing APIs and large datasets is challenging and time consuming. We
> propose an alternative solution to meet this requirement while keeping it
> simple and computationally fast. We added a new argument called test.mode
> in some functions. This argument allows to run a few more lines of code to
> speed-up the testing process and avoid very specific errors that very
> difficult to anticipate (e.g., omitting calls to the API and replacing
> them with calls to GitHub mimicking the API response). In addition, we
> added a new function to test internal methods of the package, which are
> usually omitted by the test, such as functions designed to order images,
> etc... With these changes we have managed to exceed the 70% of lines
> executed in codecov.
>
> --
> You are receiving this because you were mentioned.
> Reply to this email directly or view it on GitHub:
> https://github.com/ropensci/software-review/issues/437#issuecomment-827709585
