### 2021-05-03 #437 @jhollist
@unai-perez Thanks for addressing these issues.  I think we are very close.

I have one last thing that needs to be addressed prior to starting review.  The tests, as currently constructed, don't finish on my machine.  I dug into it a little and it looks like lines 59-70 in test-download.r (https://github.com/spatialstatisticsupna/rsat/blob/381d12de6d974eee0e99e0b41f7fd2931c0b6c15/tests/testthat/test-download.R#L59-L70) are the source of the problem.  If I comment those out and re-run the tests everything completes in about 20 minutes on my machine.  These do appear to run on CI (at least GitHub Actions didn't throw any errors) so that is a little weird.  Full disclosure, I didn't spend too much time trying find the underlying issue.

I would like you and your co-authors to do the following:

1. See if you can recreate this issue with the most recent version on https://github.com/spatialstatisticsupna/rsat.  I cloned this version locally and the ran `devtools::check("rsat")`.  It ran overnight before I stopped it.  I think if you let it run an hour or so and it hasn't completed then that indicates an issue with that code.
2. If you recreate it, please try to find the issue and correct it.  If the tests run fine for you, then I am happy to move on with the review process.

Any questions or concerns about this, just let me know.
