+++ R/connections.R
@@ -70,8 +70,8 @@ connection <- setRefClass("connections",
                   "https://lpdaacsvc.cr.usgs.gov/services/inventory",
                   "earthdata"),
     scihub = api("scihub",
-                 "https://scihub.copernicus.eu",
-                 "https://scihub.copernicus.eu/apihub",
+                 "https://apihub.copernicus.eu",
+                 "https://apihub.copernicus.eu/apihub",
                  "scihub")
     # scihubs5p=api("scihubs5p",
     #               "https://s5phub.copernicus.eu",
