### 2021-05-04 #437 @jhollist
Ahh!  The fun of API packages.

Sounds like this is something that won't be a long-term issue and once fixed you can add that test back in.  I think holding that particular test back is a good short-term solution and we can move forward with the review.

Could you also please add the rOpenSci under review badge to your README?
