+++ README.Rmd
@@ -17,10 +17,7 @@ knitr::opts_chunk$set(
 
 [![CRAN version](https://www.r-pkg.org/badges/version/rsat)](https://cran.r-project.org/web/packages/rsat/) 
 [![Status at rOpenSci Software Peer Review](https://badges.ropensci.org/437_status.svg)](https://github.com/ropensci/software-review/issues/437)
-
-
 [![Lifecycle:maturing](https://img.shields.io/badge/lifecycle-maturing-blue.svg)](https://www.tidyverse.org/lifecycle/#maturing)
-
 [![Codecov test coverage](https://codecov.io/gh/spatialstatisticsupna/rsat/branch/master/graph/badge.svg)](https://codecov.io/gh/spatialstatisticsupna/rsat?branch=master)
 [![Build status](https://ci.appveyor.com/api/projects/status/2bx8qjhhk36dnkoc/branch/master?svg=true)](https://ci.appveyor.com/project/unai-perez/rsat/branch/master)
 [![R build status](https://github.com/spatialstatisticsupna/rsat/workflows/R-CMD-check/badge.svg)](https://github.com/spatialstatisticsupna/rsat/actions)
