### 2021-06-22 #437 @mhweber
@jhollist , @unai-perez,  here is my review of `rsat`.  Overall a really nice useful package and a great addition - details below.  Main issue in my review was the Earth Explorer api which still seemed to be having issues in my testing. I left two boxes unchecked below due to this api issue and suggestion of adding contributing to README.

- [ ] **Examples** (that run successfully locally) for all exported functions

---

- A useful and well-documented package
- Please add contributing guidelines to README
- Tests currently have 1 Fail and 64 Warnings:
  + Main issue appears to be the Earth Explorer api connection - tests generate HTTP 404 error.  I had same thing happen playing with my own searches using the package and my NASA Earth Explorer username and password.
  + Aside from the api connection issue, numerous Proj4 warnings in testing
  + Error I believe due to rtoi trying to use Landsat-8 with failed Earth Explorer api connection
- I like the use of band names in `rsat` rather than band number to provide generic use of unique custom functions for different satellites / missions.
- Are there **user interface improvements** that could be made?
  + It was a bit confusing in vignette how provided built-in variable such as `NDVI`, `NBR`, or `NDSI` should be used w/out a custom function, since the example only builds a custom function rather than using one of the built-ins.
  + Minor detail but it would seem more intuitive for the order of parameters in subset to be: records object, name of the slot, value of the slot, as opposed to value of slot before name of slot.
  + It would be nice to see direct example in vignette of download a simple records object, rather than just the examples shown of downloading for an rtoi or using records method on rtoi.
  + I would love to see functions prefixed with something like 'sat_' rather than use of some very generic function names within  the package such as `download`
- Warning message for `derive` has typo: message(paste0("Product not mosaicked, mosaic the product ",
                     "from you will derive variables."))
Should have 'from which you will derive'

- One suggestion that I've seen used in another package, [nhdplusTools](https://github.com/USGS-R/nhdplusTools) that would be nice for this one would be to add a section for related similar packages- there's a growing ecosystem of remote sensing R packages and would be nice to see how this package fits into / complements that ecosystem - I'm thinking of [rgee](https://github.com/r-spatial/rgee), [RSToolbox](https://bleutner.github.io/RStoolbox/), [OpenImageR](https://github.com/mlampros/OpenImageR) and others.
