### 2021-06-22 #437 @khondula
Hi @jhollist and @unai-perez , please find my review below.

- [ ] **Examples** (that run successfully locally) for all exported functions

---

**Overall comments**

This package has nice functionality to help work with satellite data. I was able to reproduce the functionality in most of the vignettes and examples, but outside of the vignettes I found it somewhat difficult to navigate the documentation. This could be improved with more adherence to ropensci's packaging guidelines about function documentation (e.g. should be more obvious to find what an rtoi object is). Second, given the numerous other packages that have overlapping functionality (listed below), the documentation should describe how this package compares whether it is compatible (e.g. can you convert between objects with class rtoi and satellite?). Additionally, there are several functions where warnings are produced based on underlying changes to rspatial packages (ie. phase out of proj4strings), which may affect functionality in the future. I'd recommend revising to address the warnings and if possible add support for terra in addition to or in favor of raster package.

**Specific comments**

* Code of conduct, contribution guidelines, and top-level documentation to return from `?rsat` are missing and suggested by rOpenSci package guidelines

* Please provide information about how `rsat` compares/relates to packages with overlapping functionality including: [RGISTools](https://github.com/spatialstatisticsupna/RGISTools), [rLandsat](https://github.com/atlanhq/rLandsat), [getSpatialData](https://jakob.schwalb-willmann.de/getSpatialData/), [getlandsat](https://docs.ropensci.org/getlandsat/), [sen2r](https://sen2r.ranghetti.info/), [MODIS](https://github.com/MatMatt/MODIS), [luna](https://github.com/rspatial/luna), [satellite](https://cran.r-project.org/web/packages/satellite/index.html), [landsat](https://cran.r-project.org/web/packages/landsat/index.html), [sits](https://github.com/e-sensing/sits)

* The log-in profiles section of the readme is somewhat confusing. The instructions recommend having the same user name and password for both services, with "username is 4 characters long and includes a period, number or underscore". I already had existing accounts with a username that did not meet this criteria, and wasn't able to use the package with those credentials. I was able to use the package by setting the credentials used in the vignette.

* **Vignette(s)** worked locally, but would be helpful to have links to rendered versions in the readme. It could also help navigate package functionality to include numbers with the logical sequence of workflow (01_search, 02_download, 03_customize, 04_process). There is also inconsistent language for these 4 verbs between vignettes, comments in documentation (readme comment `# search, acquire, customize, and process`), and function names (acquire vs download, derive vs process). Renaming and/or using object_verb format could help reduce conflicts with other packages and make package more in line with ropensci guidelines

* There are several references in vignettes where only the identifier is given not a full citation ([@ndsi2004], [@mvc1986], [@ima2019])

* The plot function at end of example in the readme did not work, seems to require adding "preview" as second argument. (The first suggestion in the error message (adding y = "view") also gave an error.)

```
> plot(navarre, "mod09ga", variable = "NDVI", breaks = seq(0, 1, 0.1))
Error in .local(x, y, ...) :
  plot needs a parameter 'y' with one of the following values: 'view', 'preview', or 'dates'.
> plot(navarre, y = "view", "mod09ga", variable = "NDVI", breaks = seq(0, 1, 0.1))
Error in .local(x, y, ...) : view mode requires product argument.
```

* function names - several functions names are in conflict with other packages: derive (ggplot2), rename (dplyr),  download (downloader). Consider renaming functions with the `object_verb` style recommended by ropensci (`rsat_download` or `rtoi_download`?)

* Majority of exported functions do not have `@return` or `@examples` in documentation, and of the examples most are set to not run. Only a few of the functions have the level of detail that allows for 'multiple points of entry' in terms of providing enough context to understand function inputs and outputs. Many of the helper functions are somewhat straightforward in terms of what they do, but it was hard to determine their utility without more explanation in the documentation. Adding top-level package documentation and potentially grouping functions (as suggested by @mhweber ) could go a long way towards improving documentation for this package.

* In addition to clarifying return value from `sat_search` in documentation, it may be helpful to return messages from `sat_search` about objects that are added to the rtoi (similar to message if there are no images resulting from the search).

* The examples included for several functions are the same workflow ending at slightly different points (derive, get_raster, get_stars, download, etc.). It would be helpful to include more than one variation of functions in the examples to demonstrate possible variations (like is done for `sat_search`). Additionally, the example workflow takes a relatively long time to run and may encounter timeout errors - is it possible to demonstrate the same functionality with fewer dates/images?

* The example for `smoothing_images` uses spplot and stack which are in packages not otherwise loaded (sp and raster). spplot produces warnings associated with upgrades in rspatial infrastructure.

```
1: In wkt(obj) : CRS object has no comment
```
* The example for `rename` doesn't run as included

* following example in sat_search doesn't run as included.

```
> navarre <- new_rtoi(
+   "Navarre", # name of the region
+   ex.navarre, # sf of the region
+   rtoi.path
+ )
Error in (function (classes, fdef, mtable)  :
  unable to find an inherited method for function ‘new_rtoi’ for signature ‘"character", "sf", "character", "missing", "missing", "missing"’
```

* warning produced for example in `mosaic`:

```
In CPL_gdalbuildvrt(source, destination, options, oo, quiet) :
  GDAL Message 1: VSIFSeekL(xxx, SEEK_END) may be really slow on GZip streams.
```

**Other comments**

* package name - The name is short and descriptive, however when searching "rsat" there are many results for "remote server administrator tools" and RSATtools, a BioConductor package for interfacing with genomics "regulatory sequence analysis tools". This package does usually come up on the first page of results if "rsat r" is specified (but not for a more generic search like "rsat help"). I think the name works, maybe there are ways for rOpenSci to help with search engine optimization.

* Tested on macOS (10.15.7). No problems and devtools::check() ran successfully without any errors/warnings/notes.
