### 2021-07-16 #437 @unai-perez
Dear @hollis

We are working on the main comments from the reviewers; however, these changes require several checks before publication as bugs may arise.  The main lines we are addressing are:

1. function renaming: these changes make it easier to improve some references in the manual but checking the whole package after every change to make sure we don't add new bugs is time consuming.
2. improving the use of the package: In this line we improve the examples in the manual and add messages in methods to give recommendations on how to use the functions.
3. comparison with other packages: as @khondula indicates we are going to add a small comparison between several packages in which we are going to omit RGISTools since it is of our authorship and we are going to discontinue this one to promote rsat.

In addition, we have detected some problems with the latest versions of R and its dependencies. For example, @khondula has detected some tests that take a long time to finish. Those examples should take only a few seconds, at least the searching commands. Some of the coding used in the modis image search, although it works correctly, should not take more than 1 second to give the result. We have detected that it is a problem that appears with the new version of R and we are solving it.

As soon as we have all these changes made and we will comment one by one all the recommendations.
