### 2021-08-06 c5ad12c
improved modis search performance
