+++ man/rsat_download.Rd
@@ -1,18 +1,18 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/download.R
-\name{download}
-\alias{download}
-\alias{download,rtoi-method}
-\alias{download,rtoi}
-\alias{download,records-method}
-\alias{download,records}
+% Please edit documentation in R/rsat_download.R
+\name{rsat_download}
+\alias{rsat_download}
+\alias{rsat_download,rtoi-method}
+\alias{rsat_download,rtoi}
+\alias{rsat_download,records-method}
+\alias{rsat_download,records}
 \title{Download the images from a \code{records} or an \code{rtoi} object}
 \usage{
-download(x, ...)
+rsat_download(x, ...)
 
-\S4method{download}{rtoi}(x, db_path, verbose = FALSE, test.mode = FALSE, ...)
+\S4method{rsat_download}{rtoi}(x, db_path, verbose = FALSE, test.mode = FALSE, ...)
 
-\S4method{download}{records}(x, out.dir, verbose = FALSE, test.mode = FALSE, ...)
+\S4method{rsat_download}{records}(x, out.dir, verbose = FALSE, test.mode = FALSE, ...)
 }
 \arguments{
 \item{x}{a \code{records} or an \code{rtoi} object.}
@@ -61,6 +61,6 @@ sat_search(
   product = "LANDSAT_TM_C1",
   dates = as.Date("1988-08-01") + seq(1, 35)
 )
-download(navarre)
+rsat_download(navarre)
 }
 }
