+++ man/rsat_mosaic.Rd
@@ -1,47 +1,38 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/mosaic.R
-\name{mosaic,rtoi,ANY-method}
-\alias{mosaic,rtoi,ANY-method}
-\alias{mosaic,records,ANY-method}
-\alias{mosaic,records}
+% Please edit documentation in R/rsat_mosaic.R
+\name{rsat_mosaic}
+\alias{rsat_mosaic}
+\alias{rsat_mosaic,rtoi-method}
+\alias{rsat_mosaic,sf,character}
 \title{Mosaic the tiles intersecting the region of interest}
 \usage{
-\S4method{mosaic}{rtoi,ANY}(x, y, ...)
+rsat_mosaic(x, ...)
 
-\S4method{mosaic}{records,ANY}(
-  x,
-  out_path,
-  db_path,
-  bfilter,
-  warp = "extent",
-  region,
-  overwrite = FALSE,
-  ...
-)
+\S4method{rsat_mosaic}{rtoi}(x, ...)
 }
 \arguments{
 \item{x}{an \code{rtoi} object.}
 
-\item{y}{omitted parameter.}
-
 \item{...}{additional arguments.}
-
-\item{out_path}{path to save the mosaicked images. By default, the path
-is defined by \code{x}.}
 
 \item{db_path}{path to the database. By default, the path
 is defined by \code{x}.}
 
+\item{region}{an sf object. Region for cropping the images around.
+By default, the path is defined by \code{x}.}
+
+\item{out_path}{path to save the mosaicked images. By default, the path
+is defined by \code{x}.}
+
 \item{bfilter}{a vector of bands to. If not supplied, all are used.}
+
+\item{overwrite}{logical argument. If \code{TRUE}, overwrites the existing
+images with the same name.}
 
 \item{warp}{character. If equal to "extent", it also crops the images
 around the \code{rtoi}. Use "" otherwise.}
 
-\item{region}{an sf object. Region for cropping the images around.
-By default, the path is defined by \code{x}.}
-
-\item{overwrite}{logical argument. If \code{TRUE}, overwrites the existing
-images with the same name.}
+\item{y}{omitted parameter.}
 }
 \description{
 Satellite measurements are divided into indivisible units called tiles. The
@@ -80,4 +71,5 @@ mosaic(navarre, overwrite = T)
 
 derive(navarre, "NDVI", product = "LANDSAT_TM_C1")
 }
+
 }

[…]
