### 2021-08-11 abdff0f
example change to use new package funcion with rsat_ prefix
