+++ DESCRIPTION
@@ -32,10 +32,7 @@ Collate:
     'cloud_mask.R'
     'connections.R'
     'data.R'
-    'derive.R'
-    'download.R'
     'list_data.R'
-    'mosaic.R'
     'mosaic_fun_SYN.R'
     'mosaic_fun_ls.R'
     'mosaic_fun_mod.R'
@@ -44,10 +41,14 @@ Collate:
     'package_tools.R'
     'plot.R'
     'preview.R'
+    'rsat-package.R'
+    'rsat_derive.R'
+    'rsat_download.R'
+    'rsat_mosaic.R'
     'search_mod.R'
     'search_ls.R'
     'search_sen.R'
-    'search.R'
+    'rsat_search.R'
     'smoothing_images.R'
     'testing_function.R'
     'variables.R'
