+++ NAMESPACE
@@ -39,6 +39,7 @@ exportMethods("[")
 exportMethods(as.data.frame)
 exportMethods(as.vector)
 exportMethods(c)
+exportMethods(dates)
 exportMethods(length)
 exportMethods(names)
 exportMethods(new_rtoi)
