### 2021-08-16 bd90b32
suppressWarnings produced by modis projection in raster package
