### 2021-08-16 63be44e
suppressWarnings produced by gdal when reading an image from zip
