### 2021-08-16 fc605db
fixed read_rtoi
