+++ man/rsat.Rd
@@ -1,0 +1,37 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/rsat.R
+\docType{package}
+\name{rsat}
+\alias{rsat}
+\title{`rsat':}
+\description{
+This package enables you searching, downloading, customizing, and processing time series
+of satellite images from Landsat, MODIS and Sentinel in a standardized way.
+}
+\details{
+El paquete crea
+
+\enumerate{
+  \item \code{mod} identifies MODIS Terra and Aqua satellite functions.
+  \item \code{sen} identifies Sentinel functions.
+  \item \code{ls7} identifies Landsat-7 functions.
+  \item \code{ls8} identifies Landsat-8 functions.
+  \item \code{ls} identifies both Landsat-7 and Landsat-8 functions.
+  \item \code{gen} identifies function for being used in any of the three platforms.
+  \item \code{var} identifies function for deriving variables in any of the three platforms.
+}
+
+Below, there is a list of the most important functions grouped by satellite
+programs, and listed in operational order. These functions include searching,
+previewing, downloading, mosaicking, deriving new variables, compositing,
+cloud masking and filling/smoothing satellite imagery.
+}
+\section{I. Landsat functions}{
+
+The Landsat program is currently releasing imagery captured by two satellites;
+the Landsat-7 and Lansat-8. The functions for both satellites are separate due
+to discrepancies in their spectral coverage and data formats. To download
+Landsat imagery with the following functions, a USGS's `EarthExplorer' account
+is required. Please, register \href{https://ers.cr.usgs.gov/register/}{here}.
+}
+
