### 2021-08-18 7619223
rsat package entry added to manual
