### 2021-08-18 3d976d7
update mosaic to use terra instead raster
