+++ man/rsat_mosaic.Rd
@@ -20,7 +20,6 @@ rsat_mosaic(x, ...)
   warp = "extent",
   region,
   overwrite = FALSE,
-  suppressWarnings = TRUE,
   ...
 )
 }
