### 2021-08-19 2630f83
updated data extraction and added get_spatraster
