### 2021-08-19 869eb31
remove get_stars entry from manual
