### 2021-08-19 d4567d3
rsat_derived with terra
