### 2021-08-23 2b15d0b
cloud mask function using terra and name change to rsat_cloudMask
