+++ man/rsat_cloudMask.Rd
@@ -1,14 +1,10 @@
 % Generated by roxygen2: do not edit by hand
 % Please edit documentation in R/cloud_mask.R
-\name{cloud_mask}
-\alias{cloud_mask}
-\alias{cloud_mask,rtoi-method}
-\alias{cloud_mask,rtoi}
+\name{rsat_cloudMask}
+\alias{rsat_cloudMask}
 \title{Create cloud mask from an rtoi}
 \usage{
-cloud_mask(x, ...)
-
-\S4method{cloud_mask}{rtoi}(x, products = "ALL", verbose = FALSE, overwrite = FALSE, ...)
+rsat_cloudMask(x, ...)
 }
 \arguments{
 \item{x}{rtoi object from which cloud masks are computed.}
