### 2021-08-23 2a3e877
updated plot to use terra and stars. Fixed some tmap messages
