+++ R/plot.R
@@ -24,7 +24,8 @@
 #' @importFrom calendR calendR
 #' @importFrom grDevices colors
 #' @importFrom sf gdal_utils
-#' @importFrom raster raster stack clamp
+#' @importFrom terra clamp
+#' @importFrom raster raster stack
 #' @importFrom stars st_apply read_stars
 #' @include rtoi.R records.R
 #' @export
@@ -40,17 +41,18 @@
 #'
 #' # path where the region is stored
 #' rtoi.path <- tempdir()
+#'
 #' # path where downloads are stored
-#' db.path <- file.path(tempdir(), "DATABASE")
+#' set_database(file.path(tempdir(), "DATABASE"))
+#'
 #' navarre <- new_rtoi(
 #'   "Navarre",
 #'   ex.navarre,
-#'   rtoi.path,
-#'   db.path
+#'   rtoi.path
 #' )
 #'
 #' # search mod09ga products
-#' sat_search(
+#' rsat_search(
 #'   region = navarre,
 #'   product = "mod09ga",
 #'   dates = as.Date("2021-03-01") + seq(1, 20)
@@ -69,20 +71,35 @@
 #' plot(navarre, "preview")
 #'
 #' # download the images
-#' download(navarre)
+#' rsat_download(navarre)
 #'
 #' #
 #' plot(navarre, "preview")
 #' # mosaic the tiles
-#' mosaic(navarre)
+#' rsat_mosaic(navarre)
 #'
-#' # plot mosaicked images
+#' # check the product mosaicked
+#' list_data(navarre)
+#'
+#' # plot mosaicked images as rgb
 #' plot(navarre, "view", product = "mod09ga")
 #'
 #' # plot with false color
 #' plot(navarre, "view",
 #'      product = "mod09ga",
 #'      band_name = c("nir", "red", "green"))
+#'
+#' # plot derived NDVI values
+#' plot(navarre,"view",
+#'      variable="NDVI",
+#'      product="mod09ga")
+#'
+#' # Set the max and min value in plot
+#' plot(navarre,"view",
+#'      variable="NDVI",
+#'      product="mod09ga",
+#'      zlim=c(0,1))
+#'
 #' }
 setMethod(
   f = "plot",
@@ -359,7 +376,7 @@ setMethod(
     }
 
     # plot
-    genPlotGIS(r = plot.list, region(x), ...)
+    genPlotGIS(r = plot.list, region(x), verbose=verbose, ...)
   }
 )
 
@@ -387,15 +404,15 @@ setMethod(
           con <- connection$getApi(get_api_name(r))
           con$pictureDownload(p.url, pre.file)
         }
-        img <- stack(pre.file)
-        extent(img) <- extent(r)
-        projection(img) <- st_crs(crs(r))$proj4string
+        img <- rast(pre.file)
+        ext(img) <- ext(r)
+        crs(img) <- crs(r)
         img.list <- c(img.list, img)
         lname <- c(lname, paste0(sat_name(r), "_", dates(r)))
       }
 
       # plot
-      genPlotGIS(r = img.list, ...)
+      genPlotGIS(r = img.list, verbose=verbose, ...)
     } else {
       message("Empty records.")
     }
@@ -434,11 +451,11 @@ read_variables <- function(zip.file, product, var.name, date, xsize, ysize) {
                        RasterIO = rasterio,
                        proxy = FALSE)
   stars.list <- do.call(c, stars.list)
-  raster.list <- as(stars.list, "Raster")
+  raster.list <- stars.list#as(stars.list, "Raster")
   names(raster.list) <- n
   return(raster.list) # TODO change to stars
 }
-#' @importFrom raster raster stretch as.matrix
+#' @importFrom terra stretch as.matrix rast
 #' @importFrom methods as
 #' @importFrom stars read_stars
[…]
