### 2021-08-23 aa231d6
fixed some errors and added ext to support terra
