+++ NEWS.md
@@ -3,6 +3,7 @@ rsat 0.1.15 (2021-08-18)
 ### NEW FEATURES
   
   * Added database path to global environment with the function `set_database("character")` and `get_database()`
+  * Added `get_SpatRaster` function to get `SpatRaster` object from `terra` package 
 
 ### MINOR IMPROVEMENTS
 
@@ -11,6 +12,8 @@ rsat 0.1.15 (2021-08-18)
   * Improved the performance of search function
   * Fixed unique in records
   * Fixed the warnings produced by mosaic and derive functions
+  * Updated `sf` objects
+  * Use `terra` package instead of `raster` 
 
 ### BUG FIXES
 
