+++ NAMESPACE
@@ -6,8 +6,8 @@ export("get_order<-")
 export("records<-")
 export("region<-")
 export(as.records)
-export(cloud_mask)
 export(dates)
+export(get_SpatRaster)
 export(get_api_name)
 export(get_database)
 export(get_dir)
@@ -24,6 +24,7 @@ export(read_rtoi)
 export(records)
 export(region)
 export(rename)
+export(rsat_cloudMask)
 export(rsat_derive)
 export(rsat_download)
 export(rsat_mosaic)
@@ -84,34 +85,15 @@ importFrom(methods,is)
 importFrom(methods,new)
 importFrom(methods,show)
 importFrom(methods,slot)
-importFrom(raster,"NAvalue<-")
 importFrom(raster,"extent<-")
-importFrom(raster,"projection<-")
-importFrom(raster,addLayer)
-importFrom(raster,aggregate)
-importFrom(raster,as.matrix)
 importFrom(raster,beginCluster)
-importFrom(raster,clamp)
 importFrom(raster,clusterR)
-importFrom(raster,crs)
 importFrom(raster,endCluster)
 importFrom(raster,extent)
-importFrom(raster,getValues)
-importFrom(raster,maxValue)
-importFrom(raster,minValue)
-importFrom(raster,mosaic)
-importFrom(raster,ncell)
-importFrom(raster,nlayers)
-importFrom(raster,predict)
-importFrom(raster,projectRaster)
 importFrom(raster,projection)
-importFrom(raster,quantile)
 importFrom(raster,raster)
 importFrom(raster,readAll)
 importFrom(raster,stack)
-importFrom(raster,stretch)
-importFrom(raster,writeRaster)
-importFrom(raster,xyFromCell)
 importFrom(rjson,fromJSON)
 importFrom(rjson,toJSON)
 importFrom(rvest,html_nodes)
@@ -129,6 +111,26 @@ importFrom(sp,proj4string)
 importFrom(stars,read_stars)
 importFrom(stars,st_apply)
 importFrom(stars,st_as_stars)
+importFrom(terra,"NAflag<-")
+importFrom(terra,"add<-")
+importFrom(terra,"ext<-")
+importFrom(terra,aggregate)
+importFrom(terra,as.matrix)
+importFrom(terra,clamp)
+importFrom(terra,crs)
+importFrom(terra,ext)
+importFrom(terra,minmax)
+importFrom(terra,mosaic)
+importFrom(terra,ncell)
+importFrom(terra,nlyr)
+importFrom(terra,predict)
+importFrom(terra,project)
+importFrom(terra,quantile)
+importFrom(terra,rast)
+importFrom(terra,stretch)
+importFrom(terra,values)
+importFrom(terra,writeRaster)
+importFrom(terra,xyFromCell)
 importFrom(tmap,tm_borders)
 importFrom(tmap,tm_compass)
 importFrom(tmap,tm_facets)
