### 2021-08-23 760b008
cloud_mask manual name change
