### 2021-08-23 eef612a
smoothing_image function name change and terra support added
