+++ man/plot-rtoi-Date-method.Rd
@@ -121,12 +121,24 @@ rsat_mosaic(navarre)
 # check the product mosaicked
 list_data(navarre)
 
-# plot mosaicked images
+# plot mosaicked images as rgb
 plot(navarre, "view", product = "mod09ga")
 
 # plot with false color
 plot(navarre, "view",
      product = "mod09ga",
      band_name = c("nir", "red", "green"))
+
+# plot derived NDVI values
+plot(navarre,"view",
+     variable="NDVI",
+     product="mod09ga")
+
+# Set the max and min value in plot
+plot(navarre,"view",
+     variable="NDVI",
+     product="mod09ga",
+     zlim=c(0,1))
+
 }
 }
