+++ man/rsat_list_data.Rd
@@ -1,14 +1,14 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/list_data.R
-\name{list_data}
-\alias{list_data}
-\alias{list_data,rtoi-method}
-\alias{list_data,rtoi}
+% Please edit documentation in R/rsat_list_data.R
+\name{rsat_list_data}
+\alias{rsat_list_data}
+\alias{rsat_list_data,rtoi-method}
+\alias{rsat_list_data,rtoi}
 \title{List the information available for an \code{rtoi}}
 \usage{
-list_data(x, ...)
+rsat_list_data(x, ...)
 
-\S4method{list_data}{rtoi}(x, ...)
+\S4method{rsat_list_data}{rtoi}(x, ...)
 }
 \arguments{
 \item{x}{an \code{rtoi} object.}
