+++ man/rsat_cloudMask.Rd
@@ -1,10 +1,14 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/cloud_mask.R
+% Please edit documentation in R/rsat_cloud_mask.R
 \name{rsat_cloudMask}
 \alias{rsat_cloudMask}
+\alias{rsat_cloudMask,rtoi-method}
+\alias{cloud_mask,rtoi}
 \title{Create cloud mask from an rtoi}
 \usage{
 rsat_cloudMask(x, ...)
+
+\S4method{rsat_cloudMask}{rtoi}(x, products = "ALL", verbose = FALSE, overwrite = FALSE, ...)
 }
 \arguments{
 \item{x}{rtoi object from which cloud masks are computed.}
