+++ man/rsat_smoothing_images.Rd
@@ -1,16 +1,16 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/smoothing_images.R
-\name{smoothing_images}
-\alias{smoothing_images}
-\alias{smoothing_images,rtoi,character-method}
+% Please edit documentation in R/rsat_smoothing_images.R
+\name{rsat_smoothing_images}
+\alias{rsat_smoothing_images}
+\alias{rsat_smoothing_images,rtoi,character-method}
 \alias{smoothing_images,rtoi,character}
-\alias{smoothing_images,RasterBrick,character-method}
-\alias{smoothing_images,RasterStack,character-method}
+\alias{rsat_smoothing_images,RasterBrick,character-method}
+\alias{rsat_smoothing_images,RasterStack,character-method}
 \title{Fill data gaps and smooth outliers in a time series of satellite images}
 \usage{
-smoothing_images(x, method, ...)
+rsat_smoothing_images(x, method, ...)
 
-\S4method{smoothing_images}{rtoi,character}(
+\S4method{rsat_smoothing_images}{rtoi,character}(
   x,
   method,
   product = "ALL",
@@ -21,9 +21,9 @@ smoothing_images(x, method, ...)
   ...
 )
 
-\S4method{smoothing_images}{RasterBrick,character}(x, method, ...)
+\S4method{rsat_smoothing_images}{RasterBrick,character}(x, method, ...)
 
-\S4method{smoothing_images}{RasterStack,character}(x, method, ...)
+\S4method{rsat_smoothing_images}{RasterStack,character}(x, method, ...)
 }
 \arguments{
 \item{x}{\code{rtoi}, \code{RasterStack} or \code{RasterBrick} containing
@@ -129,7 +129,7 @@ data(ex.ndvi.navarre)
 spplot(ex.ndvi.navarre)
 
 # smoothin and fill all the time series
-tiles.mod.ndvi.filled <- smoothing_images(ex.ndvi.navarre,
+tiles.mod.ndvi.filled <- rsat_smoothing_images(ex.ndvi.navarre,
   method = "IMA",
   only.na = TRUE
 )
