+++ NAMESPACE
@@ -16,7 +16,6 @@ export(get_order)
 export(get_preview)
 export(get_raster)
 export(get_stars)
-export(list_data)
 export(preview)
 export(print_credentials)
 export(product)
@@ -27,13 +26,14 @@ export(rename)
 export(rsat_cloudMask)
 export(rsat_derive)
 export(rsat_download)
+export(rsat_list_data)
 export(rsat_mosaic)
 export(rsat_search)
+export(rsat_smoothing_images)
 export(sat_name)
 export(set_credentials)
 export(set_database)
 export(show_variables)
-export(smoothing_images)
 export(test_function)
 exportClasses(rtoi)
 exportMethods("[")
