+++ DESCRIPTION
@@ -26,30 +26,30 @@ RoxygenNote: 7.1.1
 Collate: 
     'add2rtoi.R'
     'api.R'
-    'extent_crs.R'
-    'records.R'
-    'rtoi.R'
-    'cloud_mask.R'
     'connections.R'
     'data.R'
-    'list_data.R'
+    'extent_crs.R'
     'mosaic_fun_SYN.R'
     'mosaic_fun_ls.R'
     'mosaic_fun_mod.R'
     'mosaic_fun_sen2.R'
     'mosaic_generic.R'
     'package_tools.R'
+    'records.R'
+    'rtoi.R'
     'plot.R'
     'preview.R'
     'rsat.R'
+    'rsat_cloud_mask.R'
     'rsat_derive.R'
     'rsat_download.R'
+    'rsat_list_data.R'
     'rsat_mosaic.R'
     'search_mod.R'
     'search_ls.R'
     'search_sen.R'
     'rsat_search.R'
-    'smoothing_images.R'
+    'rsat_smoothing_images.R'
     'testing_function.R'
     'variables.R'
 VignetteBuilder: knitr
