+++ man/ex.dem.navarre.Rd
@@ -21,6 +21,6 @@ Geographically projected \code{RasterStack} with the digital elevation model
 (DEM) of the region of Navarre (Spain). The DEM was obtained from the
 \href{http://centrodedescargas.cnig.es/CentroDescargas/locale?request_locale=en}{National Center for Geographic Information}
 of Spain. The DEM is used as a covariate in the Image Mean Anomaly (IMA)
-algorithm (\code{\link{smoothing_images}}).
+algorithm (\code{\link{rsat_smoothing_images}}).
 }
 \keyword{data}
