+++ man/rsat_get_raster.Rd
@@ -1,29 +1,29 @@
 % Generated by roxygen2: do not edit by hand
 % Please edit documentation in R/rtoi.R
-\name{get_raster}
-\alias{get_raster}
-\alias{get_raster,rtoi-method}
-\alias{get_raster,rtoi}
-\alias{get_SpatRaster}
-\alias{get_SpatRaster,rtoi-method}
-\alias{get_SpatRaster,rtoi}
-\alias{get_stars}
-\alias{get_stars,rtoi-method}
-\alias{get_stars,rtoi}
+\name{rsat_get_raster}
+\alias{rsat_get_raster}
+\alias{rsat_get_raster,rtoi-method}
+\alias{rsat_get_raster,rtoi}
+\alias{rsat_get_SpatRaster}
+\alias{rsat_get_SpatRaster,rtoi-method}
+\alias{rsat_get_SpatRaster,rtoi}
+\alias{rsat_get_stars}
+\alias{rsat_get_stars,rtoi-method}
+\alias{rsat_get_stars,rtoi}
 \title{Loads into R a time series of images regarding an rtoi, satellite product,
 and remote sensing index.}
 \usage{
-get_raster(x, p, v, ...)
+rsat_get_raster(x, p, v, ...)
 
-\S4method{get_raster}{rtoi}(x, p, v, ...)
+\S4method{rsat_get_raster}{rtoi}(x, p, v, ...)
 
-get_SpatRaster(x, p, v, ...)
+rsat_get_SpatRaster(x, p, v, ...)
 
-\S4method{get_SpatRaster}{rtoi}(x, p, v, ...)
+\S4method{rsat_get_SpatRaster}{rtoi}(x, p, v, ...)
 
-get_stars(x, p, v, ...)
+rsat_get_stars(x, p, v, ...)
 
-\S4method{get_stars}{rtoi}(x, p, v, ...)
+\S4method{rsat_get_stars}{rtoi}(x, p, v, ...)
 }
 \arguments{
 \item{x}{an rtoi.}
