+++ NAMESPACE
@@ -7,15 +7,12 @@ export("records<-")
 export("region<-")
 export(as.records)
 export(dates)
-export(get_SpatRaster)
 export(get_api_name)
 export(get_database)
 export(get_dir)
 export(get_download)
 export(get_order)
 export(get_preview)
-export(get_raster)
-export(get_stars)
 export(preview)
 export(print_credentials)
 export(product)
@@ -26,6 +23,9 @@ export(rename)
 export(rsat_cloudMask)
 export(rsat_derive)
 export(rsat_download)
+export(rsat_get_SpatRaster)
+export(rsat_get_raster)
+export(rsat_get_stars)
 export(rsat_list_data)
 export(rsat_mosaic)
 export(rsat_search)
