### 2021-08-23 4e8481e
updated rsat_list_data
