+++ NAMESPACE
@@ -62,6 +62,7 @@ importFrom(curl,handle_setform)
 importFrom(curl,handle_setopt)
 importFrom(curl,new_handle)
 importFrom(fields,Tps)
+importFrom(fields,predictSE)
 importFrom(grDevices,colors)
 importFrom(httr,POST)
 importFrom(httr,add_headers)
@@ -86,9 +87,6 @@ importFrom(methods,new)
 importFrom(methods,show)
 importFrom(methods,slot)
 importFrom(raster,"extent<-")
-importFrom(raster,beginCluster)
-importFrom(raster,clusterR)
-importFrom(raster,endCluster)
 importFrom(raster,extent)
 importFrom(raster,projection)
 importFrom(raster,raster)
@@ -115,10 +113,12 @@ importFrom(terra,"NAflag<-")
 importFrom(terra,"add<-")
 importFrom(terra,"ext<-")
 importFrom(terra,aggregate)
+importFrom(terra,app)
 importFrom(terra,as.matrix)
 importFrom(terra,clamp)
 importFrom(terra,crs)
 importFrom(terra,ext)
+importFrom(terra,interpolate)
 importFrom(terra,minmax)
 importFrom(terra,mosaic)
 importFrom(terra,ncell)
