### 2021-08-24 c015434
smoothing IMA using terra
