+++ R/rsat_smoothing_images.R
@@ -32,7 +32,7 @@
 #'
 #' @references \insertRef{militino2019interpolation}{rsat}
 #'
-#' @param x \code{rtoi}, \code{RasterStack} or \code{RasterBrick} containing
+#' @param x \code{rtoi}, \code{RastespatRaster} or \code{RasterBrick} containing
 #' a time series of satellite images.
 #' @param method character argument. Defines the method used
 #' for processing the images, e.a. "IMA".
@@ -84,35 +84,36 @@
 #' \code{FALSE}  by default.
 #' }
 #'
-#' @return a \code{RasterStack} with the filled/smoothed images.
+#' @return a \code{RastespatRaster} with the filled/smoothed images.
 #'
 #' @export
-#' @importFrom fields Tps
-#' @importFrom raster beginCluster endCluster clusterR
+#' @importFrom fields Tps predictSE
 #' @importFrom terra nlyr xyFromCell values predict aggregate
-#' @importFrom terra add<- values ncell
+#' @importFrom terra add<- values ncell app interpolate
 #' @importFrom Rdpack reprompt
 #' @examples
 #' \dontrun{
-#' # load an example of NDVI time series in Navarre
+#' require(terra)
 #' data(ex.ndvi.navarre)
 #'
+#' # load an example of NDVI time series in Navarre
+#' ex.ndvi.navarre <- rast(ex.ndvi.navarre)
 #' # the raster stack with the date in julian format as name
-#' spplot(ex.ndvi.navarre)
+#' plot(ex.ndvi.navarre)
 #'
 #' # smoothin and fill all the time series
 #' tiles.mod.ndvi.filled <- rsat_smoothing_images(ex.ndvi.navarre,
-#'   method = "IMA",
-#'   only.na = TRUE
+#'   method = "IMA"
 #' )
 #' # show the filled images
-#' spplot(tiles.mod.ndvi.filled)
+#' plot(tiles.mod.ndvi.filled)
+#'
 #' # plot comparison of the cloud and the filled images
-#' tiles.mod.ndvi.comp <- stack(
+#' tiles.mod.ndvi.comp <- c(
 #'   ex.ndvi.navarre[[1]], tiles.mod.ndvi.filled[[1]],
 #'   ex.ndvi.navarre[[2]], tiles.mod.ndvi.filled[[2]]
 #' )
-#' spplot(tiles.mod.ndvi.comp, layout = c(2, 2))
+#' plot(tiles.mod.ndvi.comp)
 #' }
 setGeneric("rsat_smoothing_images", function(x,
                                         method,
@@ -131,7 +132,7 @@ setMethod("rsat_smoothing_images",
            variable = "ALL",
            test.mode = FALSE,
            ...) {
-    var_to_process <- list_data(x)
+    var_to_process <- rsat_list_data(x)
     if (!product == "ALL") {
       var_to_process <- var_to_process[var_to_process$product %in% product, ]
     }
@@ -155,11 +156,12 @@ setMethod("rsat_smoothing_images",
                                       process_folder = "IMA",
                                       ...
                                       ) {
-      process_list <- read_rtoi_dir(unlist(p), rtoi_dir)
-      rStack <- stack(process_list)
-      names(rStack) <- format(genGetDates(process_list), "%Y%j")
+      p<-unlist(p)
+      process_list <- read_rtoi_dir(p, rtoi_dir)
+      spatRaster <- rast(process_list)
+      names(spatRaster) <- format(genGetDates(process_list), "%Y%j")
       if(!test.mode)
-        genSmoothingIMA(rStack,
+        genSmoothingIMA(spatRaster,
                         AppRoot = file.path(rtoi_dir,
                                             p[1],
[…]
