### 2021-08-24 33f5df4
update rsat_smoothing_image to use terra
