### 2021-08-24 3347750
correct the messages in rsat_search
