### 2021-08-24 8a34ccd
update the manual and remove as.vector method
