+++ man/sat_name.Rd
@@ -20,3 +20,11 @@ sat_name(x)
 \description{
 Get the name of the satellite(s) from a \code{records} or an \code{rtoi}
 }
+\examples{
+# load example rtoi
+navarre <- read_rtoi(system.file("ex/Navarre",package="rsat"))
+# get the records
+rcds <- records(navarre)
+# coerce the records to dataframr
+sat_name(rcds)
+}
