+++ man/as.vector-records-method.Rd
@@ -1,14 +1,0 @@
-% Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/records.R
-\name{as.vector,records-method}
-\alias{as.vector,records-method}
-\title{Create vector}
-\usage{
-\S4method{as.vector}{records}(x)
-}
-\arguments{
-\item{x}{records object.}
-}
-\description{
-Create a vector from \code{x}.
-}
