+++ NAMESPACE
@@ -38,7 +38,6 @@ export(test_function)
 exportClasses(rtoi)
 exportMethods("[")
 exportMethods(as.data.frame)
-exportMethods(as.vector)
 exportMethods(c)
 exportMethods(dates)
 exportMethods(length)
