### 2021-08-24 7dd99d1
fixed empty rtoi run
