+++ man/get-set_database.Rd
@@ -1,0 +1,51 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/rtoi.R
+\name{get_database}
+\alias{get_database}
+\alias{get_database,rtoi-method}
+\alias{get_database,rtoi}
+\alias{get_database,missing-method}
+\alias{set_database}
+\alias{set_database,rtoi-method}
+\alias{set_database,character-method}
+\title{Extracts or assign the path of the database}
+\usage{
+get_database(x)
+
+\S4method{get_database}{rtoi}(x)
+
+\S4method{get_database}{missing}()
+
+set_database(x, ...)
+
+\S4method{set_database}{rtoi}(x, value)
+
+\S4method{set_database}{character}(x)
+}
+\arguments{
+\item{x}{an rtoi object.}
+
+\item{value}{character argument. The value for
+change the database directory of x.}
+}
+\description{
+Extracts the path to the database from an rtoi/package environment.
+If both, environment and rtoi database are defined the rtoi
+database is used.
+}
+\examples{
+# load example rtoi
+navarre <- read_rtoi(system.file("ex/Navarre",package="rsat"))
+
+# get the databse used by navarre
+get_database(navarre)
+
+# set the a new database path
+set_database(navarre,"new_path")
+
+# get the databse used by navarre
+get_database()
+
+# set the a new database path for the entire environment
+set_database("new_path")
+}
