### 2021-08-25 d5f0169
fixed write rtoi
