### 2021-08-25 85f69ee
fixed read_rtoi
