+++ NAMESPACE
@@ -13,6 +13,7 @@ export(get_dir)
 export(get_download)
 export(get_order)
 export(get_preview)
+export(new_record)
 export(preview)
 export(print_credentials)
 export(product)
@@ -97,6 +98,7 @@ importFrom(rvest,html_nodes)
 importFrom(sf,"st_crs<-")
 importFrom(sf,gdal_crs)
 importFrom(sf,gdal_utils)
+importFrom(sf,read_sf)
 importFrom(sf,st_as_sf)
 importFrom(sf,st_bbox)
 importFrom(sf,st_crs)

[…]
