+++ man/rsat_preview.Rd
@@ -1,24 +1,24 @@
 % Generated by roxygen2: do not edit by hand
 % Please edit documentation in R/preview.R
-\name{preview}
-\alias{preview}
-\alias{preview,rtoi,Date-method}
-\alias{preview,rtoi,date}
-\alias{preview,rtoi,missing-method}
-\alias{preview,rtoi,missing}
-\alias{preview,records,Date-method}
-\alias{preview,records,date}
-\alias{preview,records,numeric-method}
-\alias{preview,rtoi,numeric}
+\name{rsat_preview}
+\alias{rsat_preview}
+\alias{rsat_preview,rtoi,Date-method}
+\alias{rsat_preview,rtoi,date}
+\alias{rsat_preview,rtoi,missing-method}
+\alias{rsat_preview,rtoi,missing}
+\alias{rsat_preview,records,Date-method}
+\alias{rsat_preview,records,date}
+\alias{rsat_preview,records,numeric-method}
+\alias{rsat_preview,rtoi,numeric}
 \title{Preview a \code{records} or an \code{rtoi} object}
 \usage{
-preview(x, n, ...)
+rsat_preview(x, n, ...)
 
-\S4method{preview}{rtoi,Date}(x, n, lpos = c(3, 2, 1), add.layer = FALSE, verbose = FALSE, ...)
+\S4method{rsat_preview}{rtoi,Date}(x, n, lpos = c(3, 2, 1), add.layer = FALSE, verbose = FALSE, ...)
 
-\S4method{preview}{rtoi,missing}(x, n, lpos = c(3, 2, 1), add.layer = FALSE, verbose = FALSE, ...)
+\S4method{rsat_preview}{rtoi,missing}(x, n, lpos = c(3, 2, 1), add.layer = FALSE, verbose = FALSE, ...)
 
-\S4method{preview}{records,Date}(
+\S4method{rsat_preview}{records,Date}(
   x,
   n,
   lpos = c(3, 2, 1),
@@ -29,7 +29,7 @@ preview(x, n, ...)
   ...
 )
 
-\S4method{preview}{records,numeric}(
+\S4method{rsat_preview}{records,numeric}(
   x,
   n,
   lpos = c(3, 2, 1),
@@ -69,38 +69,22 @@ Preview a \code{records} or an \code{rtoi} object
 \dontrun{
 library(rsat)
 
-# load navarre sf from the package
-data(ex.navarre)
+# load example rtoi
+navarre <- read_rtoi(system.file("ex/Navarre",package="rsat"))
 
-# set the credentials
 set_credentials("username", "password")
+set_database(file.path(tempdir(), "DATABASE"))
 
-# path where the region is stored
-rtoi.path <- tempdir()
-# path where downloads are stored
-db.path <- file.path(tempdir(), "DATABASE")
-navarre <- new_rtoi(
-  "Navarre",
-  ex.navarre,
-  rtoi.path,
-  db.path
-) #'
-
-sat_search(
-  region = navarre,
-  product = "S2MSI1C",
-  dates = as.Date("2018-01-01") + seq(1, 30, 1)
-)
-
-preview(navarre)
+# by default the first date in rtoi is previewed
+rsat_preview(navarre)
 
 
-sat_search(
-  region = ex.navarre,
-  product = "S2MSI1C",
-  dates = as.Date("2018-01-01") + seq(1, 30, 1)
-)
+preview.dates <- dates(navarre)
+# use add.layer to preview images of several days
+rsat_preview(navarre,preview.dates[2],add.layer = TRUE)
 
-preview(navarre, n = 1)
+# you can also preview records
+rcrds <- records(navarre)
+rsat_preview(rcrds, n = 1)
 }
 }
