+++ NAMESPACE
@@ -14,7 +14,6 @@ export(get_download)
 export(get_order)
 export(get_preview)
 export(new_record)
-export(preview)
 export(print_credentials)
 export(product)
 export(read_rtoi)
@@ -29,6 +28,7 @@ export(rsat_get_raster)
 export(rsat_get_stars)
 export(rsat_list_data)
 export(rsat_mosaic)
+export(rsat_preview)
 export(rsat_search)
 export(rsat_smoothing_images)
 export(sat_name)
@@ -112,6 +112,7 @@ importFrom(stars,st_apply)
 importFrom(stars,st_as_stars)
 importFrom(terra,"NAflag<-")
 importFrom(terra,"add<-")
+importFrom(terra,"crs<-")
 importFrom(terra,"ext<-")
 importFrom(terra,aggregate)
 importFrom(terra,app)
