+++ DESCRIPTION
@@ -38,13 +38,13 @@ Collate:
     'records.R'
     'rtoi.R'
     'plot.R'
-    'preview.R'
     'rsat.R'
     'rsat_cloud_mask.R'
     'rsat_derive.R'
     'rsat_download.R'
     'rsat_list_data.R'
     'rsat_mosaic.R'
+    'rsat_preview.R'
     'search_mod.R'
     'search_ls.R'
     'search_sen.R'
