+++ man/rsat_preview.Rd
@@ -1,5 +1,5 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/preview.R
+% Please edit documentation in R/rsat_preview.R
 \name{rsat_preview}
 \alias{rsat_preview}
 \alias{rsat_preview,rtoi,Date-method}
