+++ R/records.R
@@ -29,25 +29,25 @@
 #' @include extent_crs.R
 #' @examples
 #' \dontrun{
-#' data(ex.navarre)
-#' # Create a records of Sentinel-2 Level 1C images
-#' s2.lvl1.result <- sat_search(
-#'   region = ex.navarre,
-#'   product = "S2MSI1C",
-#'   dates = as.Date("2018-01-01") + seq(1, 30, 1)
-#' )
+#' library(rsat)
+#' # create a copy of navarre
+#' file.copy(from=system.file("ex/Navarre",package="rsat"),
+#'          to=tempdir(),
+#'          recursive = TRUE)
 #'
-#' # Create a records of Sentinel-2 Level 2A images
-#' s2.lvl2.result <- sat_search(
-#'   region = ex.navarre,
-#'   product = "S2MSI2A",
-#'   dates = as.Date("2019-01-01") + seq(1, 30, 1)
-#' )
-#' class(s2.lvl2.result)
-#' dates(s2.lvl2.result)
-#' all.records <- c(s2.lvl1.result, s2.lvl2.result)
+#' # load example rtoi
+#' navarre <- read_rtoi(file.path(tempdir(),"Navarre"))
 #'
-#' print(all.records)
+#' rcrds <- records(navarre)
+#'
+#' modis.rcrds <- rcrds[sat_name(rcrds)%in%"Modis"]
+#' ls8.rcrds <- rcrds[sat_name(rcrds)%in%"Landsat-8"]
+#'
+#' class(modis.rcrds)
+#' dates(ls8.rcrds)
+#' modis.ls8.records <- c(ls8.rcrds, modis.rcrds)
+#'
+#' print(modis.ls8.records)
 #' }
 setClass(
   "records",
