+++ tests/testthat/test-records.R
@@ -26,9 +26,9 @@ test_that("records test", {
   print(rcds)
   print(rcds[1])
 
-  rcds1 <- subset(rcds, unique(product(rcds))[1], "product")[1]
-  rcds2 <- subset(rcds, unique(product(rcds))[2], "product")[1]
-  rcds3 <- subset(rcds, unique(product(rcds))[3], "product")[1]
+  rcds1 <- subset(rcds, "product", unique(product(rcds))[1])[1]
+  rcds2 <- subset(rcds, "product", unique(product(rcds))[2])[1]
+  rcds3 <- subset(rcds, "product", unique(product(rcds))[3])[1]
   rcds.subset <- c(rcds1, rcds2, rcds3)
   rcds.subset@preview<-paste0(
     "https://unai-perez.github.io/rsat-test/previewImg/",
