+++ R/rsat_smoothing_images.R
@@ -92,7 +92,32 @@
 #' @importFrom terra add<- values ncell app interpolate
 #' @importFrom Rdpack reprompt
 #' @examples
-#' \dontrun{
+#' ## Smooth data in rtoi
+#' library(rsat)
+#' require(terra)
+#'
+#' # create a copy of pamplona in temp file
+#' file.copy(from=system.file("ex/PamplonaDerived",package="rsat"),
+#'          to=tempdir(),
+#'          recursive = TRUE)
+#'
+#' # load example rtoi
+#' pamplona <- read_rtoi(file.path(tempdir(),"PamplonaDerived"))
+#' rsat_smoothing_images(pamplona,
+#'                       method = "IMA",
+#'                       variable="NDVI"
+#' )
+#' rsat_list_data(pamplona)
+#' # get smoothed
+#' smoothed <- rsat_get_SpatRaster(pamplona,p="mod09ga",v="NDVI",s="IMA")
+#' plot(smoothed)
+#'
+#' # get original
+#' original <- rsat_get_SpatRaster(pamplona,p="mod09ga",v="NDVI",s="variables")
+#' plot(original)
+#' plot(smoothed[[1]]-original[[1]])
+#'
+#' ## smooth user defined SpatRaster dataset
 #' require(terra)
 #' data(ex.ndvi.navarre)
 #'
@@ -103,9 +128,7 @@
 #'
 #' # smoothin and fill all the time series
 #' tiles.mod.ndvi.filled <- rsat_smoothing_images(ex.ndvi.navarre,
-#'   method = "IMA",
-#'   predictSE =TRUE
-#'
+#'   method = "IMA"
 #' )
 #' # show the filled images
 #' plot(tiles.mod.ndvi.filled)
@@ -116,7 +139,6 @@
 #'   ex.ndvi.navarre[[2]], tiles.mod.ndvi.filled[[2]]
 #' )
 #' plot(tiles.mod.ndvi.comp)
-#' }
 setGeneric("rsat_smoothing_images", function(x,
                                         method,
                                         ...) {
