+++ tests/testthat/test-download.R
@@ -32,26 +32,26 @@ test_that("download test", {
 
   set_credentials("rsat.package", "UpnaSSG.2021")
   tryCatch({
-    sat_search(
+    rsat_search(
       region = navarre,
       product = c("mod09ga"),
       dates = as.Date("2021-03-01") + seq(1, 2),
       verbose=TRUE
     )
-    download(navarre,test.mode=T)
+    rsat_download(navarre,test.mode=T)
   }, error = function(e) {
     print(e)
   })
 
   tryCatch({
-    r<-sat_search(
+    r<-rsat_search(
       region = ex.navarre,
       product = c("LANDSAT_8_C1"),
       dates = as.Date("2021-03-01") + seq(1, 20),
       verbose=TRUE
     )
     records(navarre)<-c(records(navarre),r[1])
-    download(r[1],out.dir=file.path(tempdir(),"Database"))
+    rsat_download(r[1],out.dir=file.path(tempdir(),"Database"))
   }, error = function(e) {
     print(e)
   })
@@ -70,45 +70,45 @@ test_that("download test", {
   # })
 
   records(navarre)<-rcds
-  download(rcds[2:3],out.dir=file.path(tempdir(),"Database"),test.mode=T)
+  rsat_download(rcds[2:3],out.dir=file.path(tempdir(),"Database"),test.mode=T)
   #records(navarre)<-rcds
 
   tryCatch({
-    mosaic(navarre)
+    rsat_mosaic(navarre)
     plot(navarre, "view", product = unique(product(navarre))[1])
   }, error = function(e) {
-    mosaic(navarre)
+    rsat_mosaic(navarre)
   })
   plot(navarre, "view", product = unique(product(navarre))[2])
   plot(navarre, "view", product = unique(product(navarre))[3])
 
   print(navarre)
   tryCatch({
-    list_data(navarre)
+    rsat_list_data(navarre)
   }, error = function(e) {
   })
 
-  derive(navarre,product="LANDSAT_8_C1_lvl2",variable="NDVI")
-  derive(navarre,product="mod09ga",variable="NDVI")
+  rsat_derive(navarre,product="LANDSAT_8_C1_lvl2",variable="NDVI")
+  rsat_derive(navarre,product="mod09ga",variable="NDVI")
 
-  derive(navarre, "NDVI", product = "S2MSI2A",fun= function(red, blue) {
+  rsat_derive(navarre, "NDVI", product = "S2MSI2A",fun= function(red, blue) {
     ndvi <- (blue - red) / (blue + red)
     return(ndvi)
   })
 
 
   tryCatch({
-    list_data(navarre)[list_data(navarre)$variable=="NDVI",]
+    rsat_list_data(navarre)[rsat_list_data(navarre)$variable=="NDVI",]
   }, error = function(e) {
     print(e)
   })
 
   tryCatch({
-    derive(navarre,product="mod09ga",variable="NDVI")
+    rsat_derive(navarre,product="mod09ga",variable="NDVI")
     suppressWarnings(plot(navarre,"view",
                           variable="NDVI",
                           product = "mod09ga"))
-    suppressWarnings(smoothing_images(navarre,"IMA",
+    suppressWarnings(rsat_smoothing_images(navarre,"IMA",
                                       variable="NDVI",
                                       product = "mod09ga",
                                       test.mode=TRUE))
@@ -116,29 +116,29 @@ test_that("download test", {
     print(e)
   })
 
[…]

[…]
