+++ README.Rmd
@@ -38,7 +38,12 @@ You can install the development version from [GitHub](https://github.com/) with:
 
 ``` r
 # install.packages("devtools")
-devtools::install_github("spatialstatisticsupna/rsat", build_vignettes=TRUE)
+# install the package
+devtools::install_github("spatialstatisticsupna/rsat")
+
+# you can install the vignette in rsat using the following commands
+# install.packages(rmarkdown)
+# devtools::install_github("spatialstatisticsupna/rsat", build_vignettes=TRUE)
 ```
 
 ### Linux
