### 2021-09-01 b4ab910
manual and readme update
