+++ vignettes/rsat4_process.Rmd
@@ -1,5 +1,5 @@
 ---
-title: "process"
+title: "rsat4_process"
 output: rmarkdown::html_vignette
 vignette: >
   %\VignetteIndexEntry{process}
