+++ vignettes/rsat1_search.Rmd
@@ -1,5 +1,5 @@
 ---
-title: "search"
+title: "rsat1_search"
 output: rmarkdown::html_vignette
 vignette: >
   %\VignetteIndexEntry{search}
