+++ inst/REFERENCES.bib
@@ -86,3 +86,23 @@
   year={1988},
   publisher={Elsevier}
 }
+@article{holben1986characteristics,
+  title={Characteristics of maximum-value composite images from temporal AVHRR data},
+  author={Holben, Brent N},
+  journal={International journal of remote sensing},
+  volume={7},
+  number={11},
+  pages={1417--1434},
+  year={1986},
+  publisher={Taylor \& Francis}
+}
+@article{salomonson2004estimating,
+  title={Estimating fractional snow cover from MODIS using the normalized difference snow index},
+  author={Salomonson, Vincent V and Appel, I},
+  journal={Remote sensing of environment},
+  volume={89},
+  number={3},
+  pages={351--360},
+  year={2004},
+  publisher={Elsevier}
+}
