+++ README.md
@@ -1,24 +1,37 @@
+---
+editor_options: 
+  markdown: 
+    wrap: 72
+---
 
 # rsat
 
 <!-- badges: start -->
 
-[![CRAN version](https://www.r-pkg.org/badges/version/rsat)](https://cran.r-project.org/web/packages/rsat/)
-[![Status at rOpenSci Software Peer Review](https://badges.ropensci.org/437_status.svg)](https://github.com/ropensci/software-review/issues/437)
+[![CRAN
+version](https://www.r-pkg.org/badges/version/rsat)](https://cran.r-project.org/web/packages/rsat/)
+[![Status at rOpenSci Software Peer
+Review](https://badges.ropensci.org/437_status.svg)](https://github.com/ropensci/software-review/issues/437)
 [![Lifecycle:maturing](https://img.shields.io/badge/lifecycle-maturing-blue.svg)](https://www.tidyverse.org/lifecycle/#maturing)
-[![Codecov test coverage](https://codecov.io/gh/spatialstatisticsupna/rsat/branch/master/graph/badge.svg)](https://codecov.io/gh/spatialstatisticsupna/rsat?branch=master)
-[![Build status](https://ci.appveyor.com/api/projects/status/2bx8qjhhk36dnkoc/branch/master?svg=true)](https://ci.appveyor.com/project/unai-perez/rsat/branch/master)
-[![R build status](https://github.com/spatialstatisticsupna/rsat/workflows/R-CMD-check/badge.svg)](https://github.com/spatialstatisticsupna/rsat/actions)
+[![Codecov test
+coverage](https://codecov.io/gh/spatialstatisticsupna/rsat/branch/master/graph/badge.svg)](https://codecov.io/gh/spatialstatisticsupna/rsat?branch=master)
+[![Build
+status](https://ci.appveyor.com/api/projects/status/2bx8qjhhk36dnkoc/branch/master?svg=true)](https://ci.appveyor.com/project/unai-perez/rsat/branch/master)
+[![R build
+status](https://github.com/spatialstatisticsupna/rsat/workflows/R-CMD-check/badge.svg)](https://github.com/spatialstatisticsupna/rsat/actions)
+
 <!-- badges: end -->
 
 The goal of `rsat` is to help you handling time-series of satellite
 images from multiple platforms in a local, efficient, and standardized
 way. The package provides tools to;
 
-1.  Search
-2.  Download
-3.  Customize, and
-4.  Process
+1.  Search (run `vignette("rsat1_search", package = "rsat")` command)
+2.  Download (run `vignette("rsat2_download", package = "rsat")`
+    command)
+3.  Customize, and (run `vignette("rsat3_customize", package = "rsat")`
+    command)
+4.  Process (run `vignette("rsat4_process", package = "rsat")` command)
 
 satellite images from Landsat, MODIS, and Sentinel for a region and time
 of interest.
@@ -61,11 +74,11 @@ The registration in the following online portals is required to get a
 full access to satellite images with `rsat`;
 
 -   [EarthData](https://ers.cr.usgs.gov/register/): A repository of
-    NASA’s earth observation data-sets. More information about EarthData
+    NASA's earth observation data-sets. More information about EarthData
     can be found
     [here](https://earthdata.nasa.gov/earth-observation-data).
 -   [SciHub](https://scihub.copernicus.eu/dhus/#/self-registration), a
-    web service giving access to Copernicus’ scientific data hub. Please
+    web service giving access to Copernicus' scientific data hub. Please
     go [here](https://scihub.copernicus.eu/) to find more details about
     the data hub.
 
@@ -112,12 +125,94 @@ plot(navarre, "view" ,
[…]
