+++ README.Rmd
@@ -25,10 +25,10 @@ knitr::opts_chunk$set(
 
 The goal of `rsat` is to help you handling time-series of satellite images from multiple platforms in a local, efficient, and standardized way. The package provides tools to;
 
-1.  Search
-2.  Download
-3.  Customize, and
-4.  Process
+1.  Search (run `vignette("rsat1_search", package = "rsat")`)
+2.  Download (run `vignette("rsat2_download", package = "rsat")`)
+3.  Customize, and (run `vignette("rsat3_customize", package = "rsat")`)
+4.  Process (run `vignette("rsat4_process", package = "rsat")`)
 
 satellite images from Landsat, MODIS, and Sentinel for a region and time of interest.
 
@@ -111,6 +111,89 @@ See the vignettes for more examples:
 ```
 browseVignettes("rsat")
 ```
+## Related similar packages
+
+R has become an outstanding tool for remote sensing image analysis.
+There are several tools for the search and acquisition of satellite
+images, however, rsat is the first package that standardizes all the
+procedures in data acquisition to provide an unique workflow for any
+multispectral satellite.
+
+Currently there are several packages dedicated to remote sensing topic,
+but they are usually ad-hoc packages for each satellite. Here is a list
+of some of the most popular R packages dedicated to satellite imagery:
+
+### Multi satellite packages
+
+-   [RGISTools](https://github.com/spatialstatisticsupna/RGISTools)
+
+-   [getSpatialData](https://github.com/16EAGLE/getSpatialData)
+
+-   [luna](https://github.com/rspatial/luna)
+
+The closest package to `rsat` is RGISTools. `rsat` is the redefinition
+of the RGISTools package reprogrammed from scrach in the object-oriented
+programming paradigm. Many of the RGISTools code lines have been used to
+develop `rsat`, but these have been optimized and redundancies in the
+code have been removed in order to facilitate its maintenance. In
+addition, `rsat` contains new features and R classes to make it more
+user-friendly.
+
+`getSpatialData` is another package very similar to `rsat`. The package
+has the same philosophy of having a single package for searching and
+downloading satellite images. However, the development of `rsat` goes a
+bit further and in addition to search and download, the package helps
+you to organize all the downloaded information in a structured database.
+`rsat` allows you to use the metadata of the images to see the direct
+relation with your region of interest before downloading it. Also all
+image processing standardization is not developed in `getSpatialData`.
+
+The last package dedicated to image downloading is `luna`.Searching and
+downloading images compared to rsat is a bit more complicated. It is
+only able to search and download Modis and Landsat images, and does not
+help you in organizing the image products.
+
+### Single satellite packages
+
+-   [rLandsat](https://github.com/atlanhq/rLandsat)
+
+-   [getLandsat](https://github.com/ropensci/getlandsat)
+
+-   [sen2r](https://github.com/ranghetti/sen2r)
+
+`rLandsat` makes it easy to search for Landsat8 product IDs, place an
+order on USGS-ESPA and download the data. `rsat` on the other hand is
+able to do the image search without knowing all the ids, just using a
+polygon of the region of interest, making the search process much
+easier.
+
+`getlandsat` provides access to Landsat 8 metadata and images hosted on
+AWS S3 at. The package only data for the users, and does not help in
+further use, as rsat does.
+
+`sen2r` is an R library which helps to download and preprocess
+Sentinel-2 optical images. This is done through a GUI, something that
+can be very interesting for users but limits the analysis of the
+information prior to downloading, which can be done with `rsat`.
+
+### Raster processing packages
+
+-   [landsat](https://cran.r-project.org/web/packages/landsat/index.html)
+
[…]
