+++ README.md
@@ -1,45 +1,23 @@
----
-editor_options: 
-  markdown: 
-    wrap: 72
----
-
 # rsat
 
 <!-- badges: start -->
 
-[![CRAN
-version](https://www.r-pkg.org/badges/version/rsat)](https://cran.r-project.org/web/packages/rsat/)
-[![Status at rOpenSci Software Peer
-Review](https://badges.ropensci.org/437_status.svg)](https://github.com/ropensci/software-review/issues/437)
-[![Lifecycle:maturing](https://img.shields.io/badge/lifecycle-maturing-blue.svg)](https://www.tidyverse.org/lifecycle/#maturing)
-[![Codecov test
-coverage](https://codecov.io/gh/spatialstatisticsupna/rsat/branch/master/graph/badge.svg)](https://codecov.io/gh/spatialstatisticsupna/rsat?branch=master)
-[![Build
-status](https://ci.appveyor.com/api/projects/status/2bx8qjhhk36dnkoc/branch/master?svg=true)](https://ci.appveyor.com/project/unai-perez/rsat/branch/master)
-[![R build
-status](https://github.com/spatialstatisticsupna/rsat/workflows/R-CMD-check/badge.svg)](https://github.com/spatialstatisticsupna/rsat/actions)
+[![CRAN version](https://www.r-pkg.org/badges/version/rsat)](https://cran.r-project.org/web/packages/rsat/) [![Status at rOpenSci Software Peer Review](https://badges.ropensci.org/437_status.svg)](https://github.com/ropensci/software-review/issues/437) [![Lifecycle:maturing](https://img.shields.io/badge/lifecycle-maturing-blue.svg)](https://www.tidyverse.org/lifecycle/#maturing) [![Codecov test coverage](https://codecov.io/gh/spatialstatisticsupna/rsat/branch/master/graph/badge.svg)](https://codecov.io/gh/spatialstatisticsupna/rsat?branch=master) [![Build status](https://ci.appveyor.com/api/projects/status/2bx8qjhhk36dnkoc/branch/master?svg=true)](https://ci.appveyor.com/project/unai-perez/rsat/branch/master) [![R build status](https://github.com/spatialstatisticsupna/rsat/workflows/R-CMD-check/badge.svg)](https://github.com/spatialstatisticsupna/rsat/actions)
 
 <!-- badges: end -->
 
-The goal of `rsat` is to help you handling time-series of satellite
-images from multiple platforms in a local, efficient, and standardized
-way. The package provides tools to;
+The goal of `rsat` is to help you handling time-series of satellite images from multiple platforms in a local, efficient, and standardized way. The package provides tools to;
 
 1.  Search (run `vignette("rsat1_search", package = "rsat")` command)
-2.  Download (run `vignette("rsat2_download", package = "rsat")`
-    command)
-3.  Customize, and (run `vignette("rsat3_customize", package = "rsat")`
-    command)
+2.  Download (run `vignette("rsat2_download", package = "rsat")` command)
+3.  Customize, and (run `vignette("rsat3_customize", package = "rsat")` command)
 4.  Process (run `vignette("rsat4_process", package = "rsat")` command)
 
-satellite images from Landsat, MODIS, and Sentinel for a region and time
-of interest.
+satellite images from Landsat, MODIS, and Sentinel for a region and time of interest.
 
 ## Installation
 
-You can install the development version from
-[GitHub](https://github.com/) with:
+You can install the development version from [GitHub](https://github.com/) with:
 
 ``` r
 # install.packages("devtools")
@@ -53,8 +31,7 @@ devtools::install_github("spatialstatisticsupna/rsat")
 
 ### Linux
 
-In Linux, you need to install additional libraries before starting with
-`rsat`. Use the following commands for:
[…]
