+++ R/rsat.R
@@ -11,6 +11,29 @@
 #' }
 #' satellite images from Landsat, MODIS, and Sentinel for a region and time
 #' of interest.
+#' The registration in the following online portals is required to get a
+#' full access to satellite images with `rsat`;
+#' \itemize{
+#' \item \href{https://ers.cr.usgs.gov/register/}{USGS} USGS is the sole
+#' science agency for the Department of the Interior of United States.
+#' Provide access to Modis Images. More information about EarthData
+#' can be found \href{Here}{https://www.usgs.gov/}.
+#' \item \href{https://urs.earthdata.nasa.gov}{EarthData}: A repository
+#' of NASA's earth observation data-sets. More information about
+#' EarthData can be found
+#' \href{here}{https://earthdata.nasa.gov/earth-observation-data}.
+#' \item \href{https://scihub.copernicus.eu/dhus/#/self-registration}{SciHub},
+#' a web service giving access to Copernicus' scientific data hub.
+#' Please go \href{here}{https://scihub.copernicus.eu/} to find more
+#' details about the data hub.
+#' }
+#' For convenience, try to use the same username and password for all of them.
+#' To satisfy the criteria of all web services make sure that the username is 4
+#'  characters long and includes a period, number or underscore. The password must
+#'  be 12 character long and should include characters with at least one capital
+#'  letter, and numbers.
+#'
+#'
 #' @docType package
 #' @name rsat
 NULL
