### 2021-09-09 #437 @unai-perez
Dear @jhollist, @khondula and @mhweber. We have finally answered all the comments you have made. The most important and costly question has been the use of terra instead of raster. Now the raster package is only used in a few exceptional code pieces. The user manual has been substantially improved, especially the examples that now allow unitary tests without having to show the whole package workflow over and over again.
We have added a new database path management, which simplifies the rtoi and the usability of the package. The way an rtoi is stored has also been changed and improved. Some functions have better performance now and small bugs have been fixed throughout the package.
We believe that all these changes have substantially improved the package and can be better fitted in future enhancements.
In the following lines I address each comment and response to the reviewers recommendation:
**Reviewer 1**

> @jhollist , @unai-perez, here is my review of rsat. Overall a really nice useful package and a great addition - details below. Main issue in my review was the Earth Explorer api which still seemed to be having issues in my testing. I left two boxes unchecked below due to this api issue and suggestion of adding contributing to README.

Thank you very much, we think your comments have been very accurate and we have implemented all of them. We have corrected the issues and added the contributing section.

> Package Review
> Please check off boxes as applicable, and elaborate in comments below. Your review is not limited to these topics, as described in the reviewer guide
> •	Briefly describe any working relationship you have (had) with the package authors.
> •	 As the reviewer I confirm that there are no conflicts of interest for me to review this work (If you are unsure whether you are in conflict, please speak to your editor before starting your review).
> Documentation
> The package includes all the following forms of documentation:
> •   A statement of need clearly stating problems the software is designed to solve and its target audience in README
> •   Installation instructions: for the development version of package and any non-standard dependencies in README
> •   Vignette(s) demonstrating major functionality that runs successfully locally
> •   Function Documentation: for all exported functions
> •   Examples (that run successfully locally) for all exported functions
> •   Community guidelines including contribution guidelines in the README or CONTRIBUTING, and DESCRIPTION with URL, BugReports and Maintainer (which may be autogenerated via Authors@R).
> Functionality
> •   Installation: Installation succeeds as documented.
> •   Functionality: Any functional claims of the software been confirmed.
> •   Performance: Any performance claims of the software been confirmed.
> •   Automated tests: Unit tests cover essential functions of the package
> and a reasonable range of inputs and conditions. All tests pass on the local machine.
> •   Packaging guidelines: The package conforms to the rOpenSci packaging guidelines
> Estimated hours spent reviewing: 10
> •	 Should the author(s) deem it appropriate, I agree to be acknowledged as a package reviewer ("rev" role) in the package DESCRIPTION file.
> ________________________________________
> Review Comments
> •	A useful and well-documented package
> •	Please add contributing guidelines to README
> •	Tests currently have 1 Fail and 64 Warnings:
> o	Main issue appears to be the Earth Explorer api connection - tests generate HTTP 404 error. I had same thing happen playing with my own searches using the package and my NASA Earth Explorer username and password.
> o	Aside from the api connection issue, numerous Proj4 warnings in testing
> o	Error I believe due to rtoi trying to use Landsat-8 with failed Earth Explorer api connection

Fixed! this does not happen anymore

> •	I like the use of band names in rsat rather than band number to provide generic use of unique custom functions for different satellites / missions.
> •	Are there user interface improvements that could be made?
> o	It was a bit confusing in vignette how provided built-in variable such as NDVI, NBR, or NDSI should be used w/out a custom function, since the example only builds a custom function rather than using one of the built-ins.

DONE. We have added new examples that generate different indexes. Some of them are build-in and others are user defined.

> o	Minor detail but it would seem more intuitive for the order of parameters in subset to be: records object, name of the slot, value of the slot, as opposed to value of slot before name of slot.

DONE. We have decided to change the order of these parameters.

> o	It would be nice to see direct example in vignette of download a simple records object, rather than just the examples shown of downloading for an rtoi or using records method on rtoi.

DONE. We have added an example of downloading some records.

> o	I would love to see functions prefixed with something like 'sat_' rather than use of some very generic function names within the package such as download

DONE. The most important functions in the package now are named as “rsat_”

> •	Warning message for derive has typo: message(paste0("Product not mosaicked, mosaic the product ",
> "from you will derive variables."))
> Should have 'from which you will derive'

DONE. We fixed this typo

> •	One suggestion that I've seen used in another package, nhdplusTools that would be nice for this one would be to add a section for related similar packages- there's a growing ecosystem of remote sensing R packages and would be nice to see how this package fits into / complements that ecosystem - I'm thinking of rgee, RSToolbox, OpenImageR and others.

DONE. We have added this section with a little remote sensing package comparison.

Thank you very much for your helpful comments.

>
> **Reviewer 2**
> Hi @jhollist and @unai-perez , please find my review below.
> Package Review
> Please check off boxes as applicable, and elaborate in comments below. Your review is not limited to these topics, as described in the reviewer guide
> •	Briefly describe any working relationship you have (had) with the package authors.
> •	 As the reviewer I confirm that there are no conflicts of interest for me to review this work (If you are unsure whether you are in conflict, please speak to your editor before starting your review).
> Documentation
> The package includes all the following forms of documentation:
> •   A statement of need clearly stating problems the software is designed to solve and its target audience in README
> •  Installation instructions: for the development version of package and any non-standard dependencies in README
> •  Vignette(s) demonstrating major functionality that runs successfully locally
> •  Function Documentation: for all exported functions
> •  Examples (that run successfully locally) for all exported functions
> •	   Community guidelines including contribution guidelines in the README or CONTRIBUTING, and DESCRIPTION with URL, BugReports and Maintainer (which may be autogenerated via Authors@R).
> Functionality
> •   Installation: Installation succeeds as documented.
> •  Functionality: Any functional claims of the software been confirmed.
> •  Performance: Any performance claims of the software been confirmed.
> •  Automated tests: Unit tests cover essential functions of the package
> and a reasonable range of inputs and conditions. All tests pass on the local machine.
> •   Packaging guidelines: The package conforms to the rOpenSci packaging guidelines
> Estimated hours spent reviewing:
> •	 Should the author(s) deem it appropriate, I agree to be acknowledged as a package reviewer ("rev" role) in the package DESCRIPTION file.
> ________________________________________
> Review Comments
> Overall comments
> This package has nice functionality to help work with satellite data. I was able to reproduce the functionality in most of the vignettes and examples, but outside of the vignettes I found it somewhat difficult to navigate the documentation. This could be improved with more adherence to ropensci's packaging guidelines about function documentation (e.g. should be more obvious to find what an rtoi object is). Second, given the numerous other packages that have overlapping functionality (listed below), the documentation should describe how this package compares whether it is compatible (e.g. can you convert between objects with class rtoi and satellite?). Additionally, there are several functions where warnings are produced based on underlying changes to rspatial packages (ie. phase out of proj4strings), which may affect functionality in the future. I'd recommend revising to address the warnings and if possible add support for terra in addition to or in favor of raster package.

Thank you very much for your review, we have followed your indications and we have recoded the package to use terra. We have also revised the entire manual and improved all the entries to make the package easier to use.

> Specific comments
> •	Code of conduct, contribution guidelines, and top-level documentation to return from ?rsat are missing and suggested by rOpenSci package guidelines

DONE. We added this section to the package

> •	Please provide information about how rsat compares/relates to packages with overlapping functionality including: RGISTools, rLandsat, getSpatialData, getlandsat, sen2r, MODIS, luna, satellite, landsat, sits

DONE. We added a comparison in the readme.

> •	The log-in profiles section of the readme is somewhat confusing. The instructions recommend having the same user name and password for both services, with "username is 4 characters long and includes a period, number or underscore". I already had existing accounts with a username that did not meet this criteria, and wasn't able to use the package with those credentials. I was able to use the package by setting the credentials used in the vignette.

DONE. We added how to create a user account that was missing in the readme

> •	Vignette(s) worked locally, but would be helpful to have links to rendered versions in the readme. It could also help navigate package functionality to include numbers with the logical sequence of workflow (01_search, 02_download, 03_customize, 04_process). There is also inconsistent language for these 4 verbs between vignettes, comments in documentation (readme comment # search, acquire, customize, and process), and function names (acquire vs download, derive vs process). Renaming and/or using object_verb format could help reduce conflicts with other packages and make package more in line with ropensci guidelines

DONE. We rename the vignettes and we add a link in the rsat section of the manual

> •	There are several references in vignettes where only the identifier is given not a full citation ([@ndsi2004], [@mvc1986], [@ima2019])

DONE. We fixed these references.

> •	The plot function at end of example in the readme did not work, seems to require adding "preview" as second argument. (The first suggestion in the error message (adding y = "view") also gave an error.)

DONE. We rewrite all the examples in plot to ease its usability

> > plot(navarre, "mod09ga", variable = "NDVI", breaks = seq(0, 1, 0.1))
> Error in .local(x, y, ...) :
>   plot needs a parameter 'y' with one of the following values: 'view', 'preview', or 'dates'.
> > plot(navarre, y = "view", "mod09ga", variable = "NDVI", breaks = seq(0, 1, 0.1))
> Error in .local(x, y, ...) : view mode requires product argument.
> •	function names - several functions names are in conflict with other packages: derive (ggplot2), rename (dplyr), download (downloader). Consider renaming functions with the object_verb style recommended by ropensci (rsat_download or rtoi_download?)

DONE. Now we use 'rsat_’ prefix in the main functions

> •	Majority of exported functions do not have @return or @examples in documentation, and of the examples most are set to not run. Only a few of the functions have the level of detail that allows for 'multiple points of entry' in terms of providing enough context to understand function inputs and outputs. Many of the helper functions are somewhat straightforward in terms of what they do, but it was hard to determine their utility without more explanation in the documentation. Adding top-level package documentation and potentially grouping functions (as suggested by @mhweber ) could go a long way towards improving documentation for this package.

DONE. We redefine all the examples to show the use of the package

> •	In addition to clarifying return value from sat_search in documentation, it may be helpful to return messages from sat_search about objects that are added to the rtoi (similar to message if there are no images resulting from the search).

DONE. The function now prints the number of records found in each API.

> •	The examples included for several functions are the same workflow ending at slightly different points (derive, get_raster, get_stars, download, etc.). It would be helpful to include more than one variation of functions in the examples to demonstrate possible variations (like is done for sat_search). Additionally, the example workflow takes a relatively long time to run and may encounter timeout errors - is it possible to demonstrate the same functionality with fewer dates/images?

DONE. Now the example shows more variations of these functions and we add a new function rsat_get_spatRaster to get terra raster class

> •	The example for smoothing_images uses spplot and stack which are in packages not otherwise loaded (sp and raster). spplot produces warnings associated with upgrades in rspatial infrastructure.

DONE. We redefine the function to run IMA using terra package and show its different user

> Warning messages:
> 1: In wkt(obj) : CRS object has no comment
> •	The example for rename doesn't run as included
> We have not been able to replicate the error, it could be that some library on your computer kept an image open and lock the folder name. We are following up to see if we can fix it in next commits.
> •	following example in sat_search doesn't run as included.

FIXED! this does not happen anymore

> > navarre <- new_rtoi(
> +   "Navarre", # name of the region
> +   ex.navarre, # sf of the region
> +   rtoi.path
> + )
> Error in (function (classes, fdef, mtable)  :
>   unable to find an inherited method for function ‘new_rtoi’ for signature ‘"character", "sf", "character", "missing", "missing", "missing"’
> •	warning produced for example in mosaic:
> Warning message:
> In CPL_gdalbuildvrt(source, destination, options, oo, quiet) :
>   GDAL Message 1: VSIFSeekL(xxx, SEEK_END) may be really slow on GZip streams.

FIXED! this does not happen anymore

> Other comments
> •	package name - The name is short and descriptive, however when searching "rsat" there are many results for "remote server administrator tools" and RSATtools, a BioConductor package for interfacing with genomics "regulatory sequence analysis tools". This package does usually come up on the first page of results if "rsat r" is specified (but not for a more generic search like "rsat help"). I think the name works, maybe there are ways for rOpenSci to help with search engine optimization.
> •	Tested on macOS (10.15.7). No problems and devtools::check() ran successfully without any errors/warnings/notes.

Thank you very much for your helpful comments.
