### 2021-09-17 a3c4718
fixed same name variable extraction
