### 2021-09-29 #437 @khondula
Wow, impressive work @unai-perez!
I'm afraid I'm missing something - I am no longer able to run the vignettes and can't figure out why?

> library(rsat)
> browseVignettes(package = "rsat")
No vignettes found by browseVignettes(package = "rsat")
> vignette("rsat1_search", package = "rsat")
Warning message:
vignette ‘rsat1_search’ not found
