### 2021-09-30 af2b161
update installation guide
