+++ R/connections.R
@@ -60,7 +60,7 @@ connection <- setRefClass("connections",
     eeapi = api("earthexplorer",
                 "https://earthexplorer.usgs.gov",
                 "https://m2m.cr.usgs.gov/api/api/json/stable",
-                "earthdata"),
+                "usgs"),
     espa = api("ESPA",
                "https://espa.cr.usgs.gov",
                "https://espa.cr.usgs.gov/api/v1",
@@ -68,7 +68,7 @@ connection <- setRefClass("connections",
     nasainv = api("nasa_inventory",
                   "https://lpdaacsvc.cr.usgs.gov",
                   "https://lpdaacsvc.cr.usgs.gov/services/inventory",
-                  "earthdata"),
+                  "usgs"),
     scihub = api("scihub",
                  "https://apihub.copernicus.eu",
                  "https://apihub.copernicus.eu/apihub",
