### 2021-09-30 #437 @jhollist
I can confirm that with `install_github("spatialstatisticsupna/rsat", build_vignettes = TRUE)` I see the 4 vignettes.
