### 2021-09-30 #437 @khondula
Thank you, that fixed it!

The example in the readme workflow produces an error when I get to the first plot - is it possible that making the toi a range of 2 dates avoids this?

```
> plot(navarre, "view" ,
+      product = "mod09ga",
+      variable = "NDVI",
+      breaks = seq(0, 1, 0.1))
Error in if (dx[i] < dd[i]) { : missing value where TRUE/FALSE needed
```

I think everything else I checked looks good. The additional documentation really improves usability!
