### 2021-09-30 #437 @jhollist
Yee haw!  This one is approved!

There are a number tasks that we need to take of now to get this "officially" approved (see https://github.com/ropensci/software-review/issues/437#issuecomment-931550150).  As this is my first time doing this, we will probably need to learn together on some of these tasks.  I'll do my best to keep the process moving along.  As you are working on these, let me know if you have any questions or concerns.

I also had a few additions for you to think about:

- [ ] I notice you are using both Appveyor and GitHub Actions.  If Appveyor is still needed, you will need to deal with this task.  If not, you can remove the Appveyor CI and rely on the GitHub Actions.
- [ ] Since you are using codecov, you will likely need to reset the webhook for that.
- [ ] I also remember that you were planning on submitting your package to CRAN.  If that still is the plan, check out the [ROpenSci CRAN Gotchas](https://devguide.ropensci.org/building.html#crangotchas) to help along with that submission.
